### 2024-05-30 #6811 @editorialbot
[PRE REVIEW]: MoSDeF-dihedral-fit: A lightweight software for fitting dihedrals within MoSDeF
**Submitting author:** @jpotoff (<a href="http://orcid.org/0000-0002-4421-8787">Jeffrey Potoff</a>)
**Repository:** https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.3
