### 2024-05-30 #6811 @kyleniemeyer
Hello @jpotoff, this pre-review issue is where we will sort out any issues, assign an editor, and find reviewers.

First, I noticed the unusual LICENSE in the software repo: https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/blob/main/LICENSE.rst. Why are you including the license information for vmd-python and VMD, when these do not appear to be distributed in this software?

Second, the paper compilation issue appears to stem from a missing comma at the end of line 6 of `paper.bib`.
