+++ paper/paper.bib
@@ -3,7 +3,7 @@
 author = {Mayne, C.G. Mayne and Saam, J, and Schulten, K. and Tajkhorshid, E. and Gumbart, J.C.},
 journal = {J. Comp. Chem.},
 volume = {34},
-issue = {32}
+issue = {32},
 pages = {2757-2770},
 title = {{Rapid parameterization of small molecules using the force field toolkit}},
 doi = {10.1002/jcc.23422},
