### 2024-05-31 #6811 @phibeck
Hi @jpotoff, thanks for your submission. I'll be looking for reviewers next. A few things you can do to speed up the review process in addition to what was pointed out above by @kyleniemeyer:

- it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @)
- there are some missing DOIs, please add them and regenerate the pdf using `@editorialbot generate pdf`
