### 2024-06-10 d28cb94
Update paper.bib
Add DOIs
