+++ paper/paper.bib
@@ -85,7 +85,6 @@ issue = {2},
 pages = {230-252},
 year = {986},
 doi = {10.1002/jcc.540070216},
-
 URL = {https://doi.org/10.1002/jcc.540070216},
 }
 
@@ -110,6 +109,7 @@ journal = {J. Amer. Chem. Soc.},
 year = {1996},
 volume = {118(45)},
 pages = {11225--11236},
+doi = {10.1021/ja9621760}
 }
 
 # Mie FF
@@ -119,7 +119,8 @@ title = {},
 journal = {Ann. Phys.},
 year = {1903},
 volume = {11},
-pages = {657}
+pages = {657},
+doi = {10.1002/andp.19033160802}
 }
 
 # CHARMM FF part 1
@@ -184,7 +185,8 @@ title = {Ueber die Anwendung des Satzes vom Virial in der kinetischen Theorie de
 journal = {Ann. d. Phys.},
 volume = {12},
 pages = {127--136},
-year = {1881}
+year = {1881},
+doi ={10.1002/andp.18812480110}
 }
 
 # Gaussian software
@@ -220,7 +222,8 @@ Title = {Open-Source Molecular Modeling Software in Chemical Engineering, with F
 journal = {AICHE J.},
 volume = {67(3)},
 pages = {e17206},
-year  = {2021}
+year  = {2021},
+doi = {10.1002/aic.17206}
 }
 
 # MoSDeF part 2
@@ -353,6 +356,7 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   year=1996,
   volume=14,
   pages={33-38},
+  doi = {10.1016/0263-7855(96)00018-5}.
   note={},
   tbstatus={Published.},
   techrep={},
@@ -372,6 +376,7 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   editor={John F. Hughes and Carlo H. S\'equin},
   pages={191-194},
   year={2001},
+  doi = {10.1145/364338.364398}.
   note={},
   tbreference={304},
   tbstatus={Published.},
@@ -393,7 +398,8 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   Volume="22",
   Number="4",
   Pages="504--506",
-  Month="Feb"
+  Month="Feb",
+  doi = {10.1093/bioinformatics/bti825}
 }
 
 
