### 2024-06-10 def78d6
Update paper.md
Remove vmd subreferences
