+++ paper/paper.md
@@ -62,7 +62,7 @@ bibliography: paper.bib
 
 Molecular Mechanics (MM) simulations (molecular dynamics and Monte Carlo) provide a third method of scientific discovery, simulation modeling, adding to the traditional theoretical and experimental scientific methods.  These molecular simulations provide visualizations and calculated properties that are difficult, too expensive, or unattainable by conventional methods.  Additionally, molecular simulations can be used to obtain insights and properties on chemicals or materials that do not currently exist, are not easily attainable, or require hard-to-achieve conditions (i.e., very high pressures and temperatures).  However, these MM models require force field parameters determined from Quantum Mechanics (QM) simulations, where the MM proper dihedrals (i.e., dihedrals) are difficult to obtain if they don't currently exist for the chosen force field. While the same QM simulations can be used to fit the dihedrals in all of the force field types, these MM dihedrals are also not easily transferable between different force fields due to the differing parameters and formulas used in each force field (e.g., Combining rules and 1-4 scaling factors).
 
-The `MoSDeF-Dihedral-Fit` [@Crawford:2023b] library lets users quickly calculate the MM proper dihedrals (dihedrals) directly from the QM simulations for several force fields (OPLS, CHARMM, TraPPE, AMBER, Mie, and Exp6) [@Jorgensen:1996; @Brooks:2009; @Lee:2016-CG; @Martin:1998; @Weiner:1984; @Weiner:1986; @Potoff2009; @Hemmen2015; @Errington1999].  The user simply has to generate or use an existing Molecular Simulation Design Framework (MoSDeF) force field XML file [@Cummings:2021; @Summers:2020; @GMSO:2019; @forcefield-utilities:2022], provide Gaussian 16 or Gaussian-style Quantum Mechanics (QM) simulation files that cover the dihedral rotation (typically, 0-360 degrees), and provide the molecular structure information in a mol2 format [@Gaussian16:2016].  The `MoSDeF-Dihedral-Fit` software uses the QM and MM data to fit the dihedral for the specific force field, fitting the constants for the OPLS dihedral equation form with the correct combining rules and 1-4 scaling factors, as specified in the MoSDeF XML force field file. This software also accounts for multiple instances of the dihedral and the molecular symmetry in the molecule, and automatically removes all the unusable cosine power series values due to this symmetry.  The user can set other dihedral energies in the molecule to zero, allowing for a more flexible and accurate dihedral fit; this allows the multiple dihedral's conformational energies to be calculated from a single dihedral angle, a strategy that was used in some of the original OPLS dihedral fits.  The `MoSDeF-Dihedral-Fit` software analytically calculates the Ryckaert-Bellemans (RB)-torsions and the periodic dihedral from the OPLS dihedral.  If another form of the dihedral equation that is not currently supported is needed, the software outputs the raw data points to enable users to fit any other dihedral form.  Therefore, the `MoSDeF-Dihedral-Fit` software allows the fitting of any dihedral form, provided the force fields and software it utilizes are supported by MoSDeF and MoSDeF-GOMC (which uses GPU Optimized Monte Carlo - GOMC) [Crawford:2023a; Crawford:2022; @Crawford:2023b; Nejahi:2019; Nejahi:2021], vmd-python [vmd-python:2016] (a derivative or modified version of the original VMD software [Humphrey:1996; Stone:2001; Eargle:2006; Stone:1998; Frishman:1995; Varshney:1994; Sanner:1995; Sharma:2000]), and the QM data is provided as a Gaussian output file, or a generalized Gaussian-style output form [Gaussian16:2016].
[…]
