### 2024-06-10 d50cb0c
Update paper.bib
Remove most of VMD subcitations
