+++ paper/paper.bib
@@ -363,7 +363,6 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   tbreference={222}
 }
 
-
 %%
 %% Interactive Molecular Dynamics, cited by papers doing IMD simulations
 %%
@@ -381,105 +380,6 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   tbreference={304},
   tbstatus={Published.},
   annote={This symposium took place between March 19-21, 2001. {ISBN} 1-58113-292-1.}
-}
-
-%%
-%% VMD-related packages and their respective citations
-%%
-
-%%
-%% Multiple Alignment Plugin, by the ZS group
-%%
-@Article{Eargle:2006,
-  Author="Eargle, John and Wright, Dan and Luthey-Schulten, Zaida",
-  Title="{Multiple Alignment of protein structures and sequences for VMD}",
-  Journal="Bioinformatics",
-  Year="2006",
-  Volume="22",
-  Number="4",
-  Pages="504--506",
-  Month="Feb",
-  doi = {10.1093/bioinformatics/bti825}
-}
-
-
-%%
-%% Tachyon ray tracing library (built-into VMD and/or distributed with VMD)
-%% Please cite Tachyon if you use it to make figures within VMD.
-%%
-@mastersthesis{Stone:1998,
-  author = {John Stone},
-  title = "{\em An Efficient Library for Parallel Ray Tracing and Animation}",
-  school = "Computer Science Department, University of Missouri-Rolla",
-  month = "April",
-  year = {1998}
-}
-
-
-%%
-%% STRIDE Secondary Structure Prediction (distributed with VMD)
-%%
-%% Frishman,D & Argos,P. (1995) Knowledge-based secondary structure
-%% assignment. Proteins: structure, function and genetics, 23, 566-579.
-
-@article{Frishman:1995,
-  author={D. Frishman and P. Argos},
-  title={Knowledge-based secondary structure assignment},
-  journal={Proteins: structure, function and genetics},
-  volume=23,
-  pages={566-579},
-  year={1995},
-  note={},
-  annote={}
-}
-
-%%
-%% SURF solvent accessible surface calculator (distributed with VMD)
-%%
-%% A. Varshney, F. P. Brooks, W. V. Wright, Linearly Scalable Computation
-%% of Smooth Molecular Surfaces, IEEE Comp. Graphics and Applications,
-%% v. 14 (1994) pp. 19-25.
-@article{Varshney:1994,
-  author={A. Varshney and F. P. Brooks and W. V. Wright},
-  title={Linearly Scalable Computation of Smooth Molecular Surfaces},
-  journal={IEEE Computer Graphics and Applications},
-  volume=14,
-  pages={19-25},
-  year={1994},
-  note={},
-  annote={}
-}
-
-%%
-%% MSMS solvent excluded surface calculator (available separately from VMD)
-%%
-@inproceedings{Sanner:1995,
-  author={Michel Sanner and Arthur Olsen and Jean-Claude Spehner},
-  title={Fast and Robust Computation of Molecular Surfaces},
-  booktitle={Proceedings of the 11th ACM Symposium on Computational Geometry},
-  address={New York},
-  publisher={ACM},
-  pages={C6-C7},
-  year={1995},
-  note={},
-  annote={}
-}
-
-
-%%
-%% Paper on speech and gesture recognition with VMD
-%%
-@article{Sharma:2000,
-  author={Rajeev Sharma and Michael Zeller and Vladimir I. Pavlovic and Thomas S. Huang and Zion Lo and Stephen Chu and Yunxin Zhao and James C. Phillips and Klaus Schulten},
[…]
