### 2024-06-10 9e1ce4a
Update paper.md
Remove most VMD subcitations
