+++ paper/paper.bib
@@ -356,7 +356,7 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   year=1996,
   volume=14,
   pages={33-38},
-  doi = {10.1016/0263-7855(96)00018-5}.
+  doi = {10.1016/0263-7855(96)00018-5},
   note={},
   tbstatus={Published.},
   techrep={},
