+++ paper/paper.bib
@@ -375,7 +375,7 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   editor={John F. Hughes and Carlo H. S\'equin},
   pages={191-194},
   year={2001},
-  doi = {10.1145/364338.364398}.
+  doi = {10.1145/364338.364398},
   note={},
   tbreference={304},
   tbstatus={Published.},
