### 2024-06-10 #6811 @phibeck
> What is the preferred method of referring to software that does not have an associated DOI?

If the software is on GitHub, referencing by URL and with project name makes sense. Seems like this is what you did already.
