### 2024-06-10 b1bbb68
Update paper.bib
Add Berthelot Reference
