+++ paper/paper.bib
@@ -164,7 +164,8 @@ title = {Sur le m\'elange des gaz},
 journal = {Comptes Rendus Hebd.~Acad.~Sci.},
 year = {1898},
 volume = {126},
-pages = {1703--1855}
+pages = {1703--1855},
+url = {https://www.biodiversitylibrary.org/item/111847#page/757/mode/1up}
 }
 
 # Mixing rule arithmetic epslion
