### 2024-06-11 #6811 @mattwthompson
I'm one of the original authors of much of the MoSDeF software stack, but haven't been involves in a few years and didn't know about this project until now. I feel I can give an impartial review (and am happy to help) if you're comfortable with that context.
