### 2024-06-11 #6811 @phibeck
Thanks @matthewthompson! Could you please let me know whether you have a COI based on the following guideline (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). It sounds like the answer is "no COI", but I just want to be sure.

> As a reviewer (or editor), COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published; and lifetime for the family members, business partners, and thesis student/advisor or mentor. In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.
