### 2024-06-14 #6811 @mattwthompson
Based on this criteria I think I would

> recent (past four years) collaborators in funded research or work that is published

I have no COI with the first or corresponding authors but I had a funded software (i.e. producing a software product, not scientific research output) project with the middle five authors in 2021-2022. That project did not produce a publication but did contribute to a software project I'm still working on. (These people are in the lab in did my PhD in, 2014-2019, including my PI Peter Cummings)

The calls is yours, of course - and in either case I'm happy advertise this to some other experts in dihedral fitting
