### 2024-07-04 #6811 @naik-aakash
Hi @phibeck, I would be happy to review this submission for the JOSS.
