### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#94 @naik-aakash
Any previous works?
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

I was reading through the paper submitted on the JOSS. There was this particular line that got me wondering if there are any previous studies that highlights this specific issue mentioned?

` Due to these limitations, scientists assumed that the dihedral fits were transferable with all the atom classes in the dihedral fit; however, this is not always an accurate assumption.  Some of the dihedrals were only fit to the first minimum and not the entire dihedral landscape, which can lead to errors in the predicted molecular conformations.`
