### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#95 @naik-aakash
Suggestion : Add coverage badge to readme
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

I was thinking, it would be great it coverage badge is added to the readme for a quick overview. Not required to be added, is just an suggestion from my end. 😄
