### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#96 @naik-aakash
Some Citations are incorrectly referred in the markdown leading to issues in the rendered paper draft
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

I noticed some citations are  incorrectly referred in the markdown , that leads to these citations being rendered as `?` in the rendered pdf version of the paper draft. (Line number 29 in the rendered PDF)

`@Potoff2009; @Hemmen2015; @Errington1999` should be `@Potoff:2009; @Hemmen:2015; @Errington:1999`

Besides these, there are some citations are not referred in the markdown without a  leading `@` , thus hyperlinks are not added for these. I have noted down the line numbers below as per PDF version of the paper draft for your reference

`Line numbers 48, 49, 50, 51, 77, 82`
