### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#97 @naik-aakash
Contributing / Community guidelines missing
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

I think it would be good to also include a contributing guide to the main repository and the rendered docs. This might encourage users who wish to contribute to this package development further easily.
