### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#98 @naik-aakash
Suggestion : Tests 
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

I feel it would be great if tests are added to check for exceptions raised as well. From the current coverage report , I see most of the missing lines are from exceptions and would be better to have tests for these lines as well.

![Screenshot 2024-07-19 200817](https://github.com/user-attachments/assets/b1809f4e-013a-4081-a0e1-504a27afae76)
