### 2024-07-20 GOMC-WSU/MoSDeF-dihedral-fit#99 @naik-aakash
Suggestions : Documentation and Tests
## Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

Dear developers,

Do not know how much of additional work it can be. I was thinking would it be feasible to add some examples for dihedral fitting using other forcefields i.e.,  CHARMM, TraPPE, AMBER, Mie, and Exp6  to the documentation ?

And also include some tests for these. I noticed MIE and OPLS are already included in tests, but could not find tests for other forcefields mentioned in the paper. (It is possible I could be missed the tests, kindly correct me if am mistaken on this.)
