### 2024-07-21 GOMC-WSU/MoSDeF-dihedral-fit#101 @bc118
Team checklist of JOSS reviewer requests
## **MOSDEF-DIHEDRAL-FIT TEAM NOTE/INSTRUCTIONS: **
#### *******************************************************************************************************

This is a task board for the mosdef-dihedral fit team, so we can track who is doing what and we are not accidentally working on the same parts.  Some of these tasks are larger and may take >1-2 weeks (I know @bc118 will), given we some are working on this in our free time, which is OK.

**Structure:**
The tasks are broken out by reviewer and their comments/questions/requests are on the line with the checkbox [ ] .

I added the following to each item to help us :
- _**mosdef-dihedral-fit team notes**_ = includes my thoughts on what to do, if any.
- _**issue number**_ so we can also write in that issue, tag it in the PR, and close it.
-  **_Person completing it:_** = so we know who is doing it, not doing work 2x.  Also, I listed the likely candidates who are best suited to do it quickly and put ????  marks there to indicate it is not decided yet.
- _**TASK COMPLETION NOTES**_  = mosdef-dihedral-fit team notes on who did it, if there is anything important that needs to be noted.  This can also be used to state as why we can or cannot do said part or whole item or what we see as a better alternative....

 needed, I am pulling a hard copy zip file of the code, as the line numbers will change as we submit the changes, so need a way of tracking them (i.e., they will be the same in the attached zip file).  I also tried to copy and paste the text that was referenced by line numbers, so it is easier for everyone.

**_Copy of MoSDeF-dihedral-fit before any changes if the line numbers are need for the review reference _**
[MoSDeF-dihedral-fit-main_pulled_7-21-2024_for_JOSS_review.zip](https://github.com/user-attachments/files/16325120/MoSDeF-dihedral-fit-main_pulled_7-21-2024_for_JOSS_review.zip)

**_Copy of MoSDeF-dihedral-fit JOSS paper f the line numbers are need for the review reference _**
[mosdef_dihedral_fit_joss.draft.pdf](https://github.com/user-attachments/files/16325832/mosdef_dihedral_fit_joss.draft.pdf)

## **MOSDEF-DIHEDRAL-FIT TEAM ** LAST TASKS AFTER ALL OTHER CODE TASKS DONE
#### *******************************************************************************************************
101.  [x] **build new conda-forge package **  @bc118
102.  [x] **Rebuild docs after all tasks completed**  -  They are auto-rebuilding
## **REVIEWERS 1's REQUESTS/SUGGESTIONS: **
#### *******************************************************************************************************

1.  [x] **_Add pytest as a dependancy to the next release version -_**  @bc118 notes: LAST CHANGE TO DO ON  CONDA-FORGE FEEDSOCK.

**mosdef-dihedral-fit team notes:**
This exists on the main 'environment.yml' and in the docs one 'docs/environment_docs.yml'.  It needs added to the conda-forge feedstock **_(add `pytest` and  `pytest-cov`)_** at the after all changes are done release script.

 issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/84

PR = https://github.com/conda-forge/mosdef-dihedral-fit-feedstock/pull/1

 Person completing it = @bc118

TASK COMPLETION NOTES = Added 'pytest', 'pytest-cov' and 'coverage' to the GitHub and conda-forge repos

 #### ----------------------------------------------------------------------------------------------------------------------------

2. [x] **_Any previous works?_** - I was reading through the paper submitted on the JOSS. There was this particular line that got me wondering if there are any previous studies that highlights this specific issue mentioned?

**mosdef-dihedral-fit team notes on the paragraph of interest:**
 Due to these limitations, scientists assumed that the dihedral fits were transferable with all the atom classes in the dihedral fit; however, this is not always an accurate assumption.  Some of the dihedrals were only fit to the first minimum and not the entire dihedral landscape, which can lead to errors in the predicted molecular conformations.

 @bc118 's comment: @jpotoff  or @chrisiacovella can you complete this??

 **If existing citations exist, lets use them.**

 _If the reviewers want to see this, I would be happy to show them though, if it provides hard evidence._

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/94

 Person completing it = @jpotoff  or @chrisiacovella can you complete this??

TASK COMPLETION NOTES =  Resolved by deleting line 74.  While this is a well-known phenomena, it is not well documented in the literature.  Most of the papers that address this phenomena are limited in scope, highlighting special cases.

#### ----------------------------------------------------------------------------------------------------------------------------

3. [x] **Suggestion : Add coverage badge to readme** -   I was thinking, it would be great it coverage badge is added to the readme for a quick overview. Not required to be added, is just an suggestion from my end.

**mosdef-dihedral-fit team notes:**
We are trying to keep these consistent across the MoSDeF ecosystem (MoSDeF-dihedral-fit, MoSDeF-GOMC, GMSO, foyer, mbuild, etc.).  Therefore, we will consider adding later when we revise all of them.

issue =  https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/95

 Person completing it = @daico007

TASK COMPLETION NOTES =

#### ----------------------------------------------------------------------------------------------------------------------------

4. [x] **Some Citations are incorrectly referred in the markdown leading to issues in the rendered paper draft** - I noticed some citations are  incorrectly referred in the markdown , that leads to these citations being rendered as `?` in the rendered pdf version of the paper draft. (Line number 29 in the rendered PDF)

`@Potoff2009; @Hemmen2015; @Errington1999` should be `@Potoff:2009; @Hemmen:2015; @Errington:1999`

Besides these, there are some citations are not referred in the markdown without a  leading `@` , thus hyperlinks are not added for these. I have noted down the line numbers below as per PDF version of the paper draft for your reference

`Line numbers 48, 49, 50, 51, 77, 82`

**mosdef-dihedral-fit team notes:**
None. We just need to fix

issue =  https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/96

 Person completing it =  @bc118

TASK COMPLETION NOTES = Completed

#### ----------------------------------------------------------------------------------------------------------------------------

5. [x] **Contributing / Community guidelines missing** - I think it would be good to also include a contributing guide to the main repository and the rendered docs. This might encourage users who wish to contribute to this package development further easily.  NOTE: Part of JOSS review : https://github.com/openjournals/joss-reviews/issues/6970

**mosdef-dihedral-fit team notes:**
None. We just need to fix

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/97   and   https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101

Update LICENSE format and add contribution guideline, as suggested in https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101 and https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/97.

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/103

 Person completing it = @daico007

TASK COMPLETION NOTES = Update LICENSE format and add contribution guideline, as suggested in https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101 and https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/97.

#### ----------------------------------------------------------------------------------------------------------------------------

6. [x] **Suggestion : Tests** - I feel it would be great if tests are added to check for exceptions raised as well. From the current coverage report , I see most of the missing lines are from exceptions and would be better to have tests for these lines as well.

**mosdef-dihedral-fit team notes:**
Let see if these are possible, and if possible fix them.  I know some checks I write are are not really possible to test, as I just write them as a catch all, to handle situations that I can not figure out how to break yet.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/98

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/106

 Person completing it = @daico007

TASK COMPLETION NOTES = Tests were created to cover more of the code. Portions of the code that are not covered, are more of catch all errors that are not known how to create the errors at this time (this protects users from something we did not expect to happen).

#### ----------------------------------------------------------------------------------------------------------------------------

7.  [x] **Suggestions : Documentation and Tests** - Do not know how much of additional work it can be. I was thinking would it be feasible to add some examples for dihedral fitting using other forcefields i.e., CHARMM, TraPPE, AMBER, Mie, and Exp6 to the documentation ?  **And also include some tests** for these. I noticed MIE and OPLS are already included in tests, but could not find tests for other forcefields mentioned in the paper. (It is possible I could be missed the tests, kindly correct me if am mistaken on this.)

**mosdef-dihedral-fit team notes:**
This will take a significant effort (bigger task/item), as we need to find all the FF parameters for the other FFs.  We will fit them as accurately as possible one of the existing QM data, noting any deviations in the FF file, making a test for each FF types.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/99

PR = https://github.com/GOMC-WSU/GOMC_Examples/pull/40

 Person completing it = @bc118

TASK COMPLETION NOTES = COMPLETED and the biggest task.  A QM simulation was setup to incorporate the other FF types.  They can be added to the examples listed and linked below and unit tests.

- Added links to new examples in the README here -  https://github.com/GOMC-WSU/MoSDeF-dihedral-fit .  There is also a link to the examples in the docs examples part.

- PR that added the examples = https://github.com/GOMC-WSU/GOMC_Examples/pull/40

- Examples here = https://github.com/GOMC-WSU/GOMC_Examples/tree/main/MoSDeF-dihedral-fit   (links in https://github.com/GOMC-WSU/MoSDeF-dihedral-fit Read me and the docs examples also)

- Completed by adding more flexibility and examples for the TraPPE, AMBER, Mie, and Exp6 PR .

- Removing the CHARMM FF from the list and as Urey-Bradley's are a lot more prevent that I thought in small examples, and MoSDeF is not supporting them at the moment.

## **REVIEWERS 2's REQUESTS/SUGGESTIONS:  **

### **_- Repository section:_**
#### xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

8. [x]  **Install Section should explain the steps, as there is an additional program installed. This is important for totally new users.**

**mosdef-dihedral-fit team notes:**
Let’s add to the docs now to download and install the 2.75a version of GOMC.

**- Add checks for GOMC software GOMC_NVT_CPU (or similarly named file) when dihedral fit function and test file run.  It will have to check the path the user provided. This should be put in the `"fit_dihedral_with_gomc"` function in the `"mosdef_dihedral_fit/dihedral_fit/fit_dihedral_with_gomc.py"` file.**
- The error check from `"mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py" `can be removed if in the main function .
- add statement in how to install GOMC in the error if not installed.
- add same installation instructions in the docs
- Make sure to add tests to cover this

- How we install in unit tests, so recommend the same procedure (who ever does this please test it)
     -  **install cmake** if not already installed
        `pip install cmake`

     -  **create directory = `<users_choice_location>` of your choice and use the same value throughout these commands.  This is where the GOMC software will be installed**:
        `mkdir <users_choice_location>`

     -  **cd to directory**:
        `cd <users_choice_location>`

     -  **clone the GOMC repositrory version 2.75a**
        `git clone https://github.com/GOMC-WSU/GOMC.git --branch v2.75a <users_choice_location>/GOMC`

     -  change permission on GOMC make file and run it
      cd <users_choice_location>/GOMC
       `chmod u+x metamake.sh `
       `./metamake.sh`

     -  OPTIONAL: Add GOMC your path so it is called automatically with path entereed as "".  To so this, you must add the following to your ".bashrc" file or equivalent file (if Mac maybe ".zshrc" depending on how you set up your specific system).  If a different than ".bashrc", replace  ".bashrc"  with ".xxxrc" file that is appropriate.  The first 2 commands add a space and title to the command so it is known and traceable later.  The commands are listed below:
        `echo "" >> ~/.bashrc`
        `echo "# Add the GOMC binary files to the path" >> ~/.bashrc`
        `echo "export PATH=<users_choice_location>/GOMC/bin${PATH:+:${PATH}}" >> ~/.bashrc`

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/107

Person completing it = @CalCraven

TASK COMPLETION NOTES = Added instructions to the docs and README to clearly install the GOMC into the environment. It is pip installed from a git repository, and the binary path is symlinked such that calls to the fitting will check the bin of the conda environment for that executable.

#### ----------------------------------------------------------------------------------------------------------------------------

9. [x]  **Suggestion: Indicate, that the box after the code example is the output.**

**mosdef-dihedral-fit team notes:**
We just need to add a note indicating where the files of interest are located at in the docs examples, if I am reading this correctly.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

Person completing it = @CalCraven

TASK COMPLETION NOTES =

completed in this task/same PR =  **Install Section should explain the steps, as there is an additional program installed. This is important for totally new users.**

Added instructions to the docs and README to clearly install the GOMC into the environment. It is pip installed from a git repository, and the binary path is symlinked such that calls to the fitting will check the bin of the conda environment for that executable.

#### ----------------------------------------------------------------------------------------------------------------------------

10. [x]  **Setup.py: Authors and Emails missing**

**mosdef-dihedral-fit team notes:**
None. Just complete it

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/104

Person completing it =  @daico007  or @CalCraven ???

TASK COMPLETION NOTES = Completed

#### ----------------------------------------------------------------------------------------------------------------------------

11. [x]  **Environment.yml: Could it be that GOMC is missing? If you state here in the docs, it is an requirement (https://mosdef-dihedral-fit.readthedocs.io/en/latest/getting_started/quick_start/quick_start.html#), then it should be part of the environment. (could also make binary detection easier)**

**mosdef-dihedral-fit team notes:**
- `Reasoning: `Yes, GOMC is missing.  We would love it to be installed as a conda package; however, it needs to be built on the unique system hardware, which generates a binary file for that hardware architecture.  It is not available or possible to put on conda-forge, or build this into the conda install functions any time in the near future.

- `Task: `Add GOMC binary detection check when running the dihedral fit function or for the entire file, maybe printing out the commands to install also.
- Let’s add to the docs now to download and install the 2.75a version of GOMC.

- Add checks for GOMC software GOMC_NVT_CPU (or similarly named file) when dihedral fit function and test file run.  It will have to check the path the user provided.
- add statement in how to install GOMC in the error if not installed.
- add same installation instructions in the docs
- Make sure to add test cases to cover this.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/107

 Person completing it = @CalCraven and @bc118

TASK COMPLETION NOTES = The install instructions now direct the user to install a source version of GOMC into the conda environment `bin`, such the executable can be found without passing any argument to the `fit` function, and it automatically will find that path.

#### ----------------------------------------------------------------------------------------------------------------------------

12. [x]  **Documentation: Repository Doc strings - API Reference Docs missing? (can be generated with `sphinx-apidoc`)**

**mosdef-dihedral-fit team notes:**
None.  Just needs completed.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

 PR   = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/112

 Person completing it =  @bc118

TASK COMPLETION NOTES = PR completed,

All the functions in the code were added to the docs, but the ones not explicitly designed for the dihedral fit users, were marked as "for advanced users".

-     Added better input tests for dihedral_angle function calculation
-
-     Added all the functions, but with warnings, saying that the inputs are not checked for correctness , and only advanced users should use them.

#### ----------------------------------------------------------------------------------------------------------------------------

13. [x]  **file_read_and_write.py: If I see it correctly this one contains unused imports.**

**mosdef-dihedral-fit team notes:**
None.  Just needs completed.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/106

Person completing it =  @daico007

TASK COMPLETION NOTES = Add more tests to cover exceptions with other misc fixes. Please refer to https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101 and https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/97.

#### ----------------------------------------------------------------------------------------------------------------------------

14. [x]  **License File: if you change the formatting, then GitHub can recognize the license file (e.g.: https://github.com/OpenFreeEnergy/kartograf/blob/main/LICENSE)**

**mosdef-dihedral-fit team notes:**
None.  Just needs completed.

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/103

 Person completing it =  @daico007  or @CalCraven ???

TASK COMPLETION NOTES = Update LICENSE format and add contribution guideline, as suggested in https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101 and https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/97.

### _**- Paper:**_
#### xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
**Introduction section (lines 12-20)**

15.  [x]  **Line 12: MD and MC are just two examples, you could indicate this with e.g.**

@bc118 found the statement = "Molecular Mechanics (MM) simulations (molecular dynamics and Monte Carlo) provide a third method of scientific discovery, simulation modeling, adding to the traditional theoretical and experimental scientific methods.    "

**mosdef-dihedral-fit team notes:**
Review and add add  as suggested to the statement .

issue = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/100

PR = https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/116

Person completing it = @bc118

TASK COMPLETION NOTES =

[…]
