### 2024-07-21 GOMC-WSU/MoSDeF-dihedral-fit#99 @bc118
Hello!

Thank you for your review and comments!  We will complete all of these that are feasible to completed. We started a issue checklist with all the reviewer comments, so we can address them and write our teams notes here https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/issues/101.

If we need any clarifications, we will kindly ask for them.  Thank you!
