### 2024-07-27 ce84c04
added Mie and Exp6 F mixing rules for non-LJ parameters in docs
added Mie and Exp6 F mixing rules for non-LJ parameters in docs
