### 2024-07-27 dd2604c
Update fit_dihedral_with_gomc.py
