### 2024-07-27 3ff10a3
[pre-commit.ci] auto fixes from pre-commit.com hooks for more information, see https://pre-commit.ci
