### 2024-07-28 4f63e68
add authors and emails to setup.py
