### 2024-07-31 6ff234a
Update CONTRIBUTING.md
Added some changes like prefer code in Issues to be text not code. and reworded 1 sentence
