### 2024-07-31 2f4f68b
Update setup.py
Changed brads email address
