### 2024-08-03 7fa97e0
Add installation instructions for GOMC in README and docs, as well as fix tests for improved error messages and automated checking of PATH for GOMC installation
