+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -1,4 +1,5 @@
 import os
+import re
 
 import numpy as np
 import pytest
@@ -11,14 +12,25 @@ from mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc import (
 from mosdef_dihedral_fit.tests.base_test import BaseTest
 
 # The below "gomc_binary_directory = "/opt/gomc_build/GOMC/bin"" is the
-# automated GitHub test binary path. It has to be fully specified, even if it is in the bashrc file.
+# automated GitHub test binary path.
 # The automated GitHub test filename and location = "MoSDeF-dihedral-fit/.github/workflows/CI.yml"
-gomc_binary_directory = "/opt/gomc_build/GOMC/bin"
-
-# User changable variable, as it needs to be run locally
+gomc_binary_directory = (
+    None  # set GOMC binary path here if you want to manually specify it
+)
+# User changable variable, if the binary is not in the environment variable "PATH"
 # Examples
 # gomc_binary_directory = "/Users/brad/Programs/GOMC/GOMC_2_75/bin"
 # gomc_binary_directory = "/Users/calcraven/Documents/Vanderbilt/Research/MoSDeF/Dihedral_Fitter/GOMC/bin"
+paths = os.environ["PATH"].split(":")
+for p in paths:
+    print(p)
+    if "GOMC/bin" in p:
+        gomc_binary_directory = p
+        break
+if not gomc_binary_directory:
+    raise OSError(
+        "Missing GOMC installation. Please install from https://github.com/GOMC-WSU/GOMC and add to shell Path, or manually set gomc_binary_directory variable in tests/test_fit_dihedral_with_gomc.py"
+    )
 
 
 # NOTE: When comparing fitted values with reference value,
@@ -4177,10 +4189,8 @@ class TestFitDihedralWithGomc(BaseTest):
     def test_gomc_binary_path_containing_the_GOMC_CPU_NVT_file_does_not_exist(
         self,
     ):
-        with pytest.raises(
-            ValueError,
-            match=r"ERROR: The 'gomc_binary_path' file does not exist or contain the GOMC 'GOMC_CPU_NVT' file.",
-        ):
+        errormsg = r"ERROR: The 'gomc_binary_path' of gomc_binary_directory does not exist or contain the GOMC 'GOMC_CPU_NVT' file. If the path is incorrect, please pass the correct absolute path for gomc_binary_path. If GOMC is not installed, go to https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/tree/main?tab=readme-ov-file#installationsetup or the Read the Docs for this repository."
+        with pytest.raises(ValueError, match=re.escape(errormsg)):
             fit_dihedral_with_gomc(
                 ["HC", "CT", "CT", "HC"],
                 self.get_fn(
