+++ mosdef_dihedral_fit/dihedral_fit/fit_dihedral_with_gomc.py
@@ -599,7 +599,9 @@ def fit_dihedral_with_gomc(
 
     if not os.path.exists(f"{gomc_binary_path}/{'GOMC_CPU_NVT'}"):
         raise ValueError(
-            f"ERROR: The 'gomc_binary_path' file does not exist or contain the GOMC 'GOMC_CPU_NVT' file."
+            f"ERROR: The 'gomc_binary_path' of {gomc_binary_path} does not exist or contain the GOMC 'GOMC_CPU_NVT' file."
+            + " If the path is incorrect, please pass the correct absolute path for gomc_binary_path."
+            + " If GOMC is not installed, go to https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/tree/main?tab=readme-ov-file#installationsetup or the Read the Docs for this repository."
         )
 
     # test the 'zero_dihedral_atom_types' input
