+++ mosdef_dihedral_fit/dihedral_fit/fit_dihedral_with_gomc.py
@@ -38,6 +38,7 @@ def fit_dihedral_with_gomc(
     gomc_cpu_cores=1,
     r_squared_min=0.98,
     r_squared_rtol=2.5e-02,
+    opls_force_k0_zero=False
 ):
     """Fit the desired dihedral to a MM force field, based on QM data.
 
@@ -67,6 +68,16 @@ def fit_dihedral_with_gomc(
     and recreated while running this function to ensure only the
     lasted data is in these folders.
 
+    NOTE: The OPLS dihedral equation
+
+    .. math::
+    opls_dihedral_n_1 &= 1/2 *(
+        &= k0 
+        &= + k1 * (1 + cos(1 * phi))
+        &= + k2 * (1 - cos(2 * phi))
+        &= + k3 * (1 + cos(3 * phi))
+        &= + k4 * (1 - cos(4 * phi))
+        &= )
 
     Parameters
     ----------
@@ -314,6 +325,21 @@ def fit_dihedral_with_gomc(
         compared to the QM data.
 
         NOTE: This value may need adjusted to get the dihedral fit to solve correctly.
+    opls_force_k0_zero: bool, default=False
+        The k0 constant is from the equation listed above. 
+        If True, this is force sets the k0 constant in the opls equation to zero (0),  
+        which is the original OPLS form. This means that the dihedral energy fit must 
+        be zero (0) at dihedral angles of (-180 and 180 degrees), which could mean 
+        the dihedral produces both positive (+) and negative energy values (-).  
+        NOTE: Using this option may not allow some dihedrals to be fit properly.
+
+        If False, the k0 constant is allowed to be fitted to a non-zero constant.  
+        This allows the k0 constant to be zero (0) or a scaler value. In this case, 
+        the dihedral's k0 constant is shifted to make the minimum of the dihedral 
+        equation equal to zero (0). This is the more standard overall dihedral form 
+        (i.e. not forcing the equation k0 constant to be zero) as it allows more 
+        dihedrals to be fit properly, and this energy shifting is in almost all newer 
+        dihedral forms now.
 
     Outputs
     -------
@@ -578,8 +604,6 @@ def fit_dihedral_with_gomc(
             else:
                 raise TypeError(print_error_value)
 
-            print("*****************")
-            print(f"value_j = {str(value_j)}")
             if isinstance(value_j, list):
                 for int_j in value_j:
                     if not isinstance(int_j, int) or int_j < 0:
@@ -715,6 +739,13 @@ def fit_dihedral_with_gomc(
         raise TypeError(
             f"ERROR: The 'qm_engine' is a {type(qm_engine)}, but it needs to be a str."
         )
+    
+    # check if opls_force_k0_zero' is a bool
+    if not isinstance(opls_force_k0_zero, bool):
+        raise TypeError(
+            "ERROR: Please enter the 'opls_force_k0_zero' file as a bool."
+        )
+
 
     # **************************************************************
     # **************************************************************
@@ -764,10 +795,8 @@ def fit_dihedral_with_gomc(
 
     # liquid_box_0_length_nm must be a value <= 999.8 nm and an interger value in angstroms <= 9998 Ang
     liquid_box_0_length_nm = 999.8
-    print(
-        "INFO: Started Building the fragment for the GOMC simulation with dihedral k values = 0"
-    )
-
+    
+    # Started Building the fragment for the GOMC simulation with dihedral k values = 0"
     box_0_liq = mb.fill_box(
         compound=[fragment],
         n_compounds=[1],
[…]
