+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -17,7 +17,7 @@ gomc_binary_directory = "/opt/gomc_build/GOMC/bin"
 
 # User changable variable, as it needs to be run locally
 # Examples
-# gomc_binary_directory = "/Users/brad/Programs/GOMC/GOMC_2_75/bin"
+gomc_binary_directory = "/home/brad/Programs/GOMC/GOMC-2.75a/bin"
 # gomc_binary_directory = "/Users/calcraven/Documents/Vanderbilt/Research/MoSDeF/Dihedral_Fitter/GOMC/bin"
 
 
@@ -50,6 +50,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=1e-03,
+            opls_force_k0_zero=True
         )
 
         assert (
@@ -191,7 +192,7 @@ class TestFitDihedralWithGomc(BaseTest):
                         "k4_kcal_per_mol",
                         "r_squared",
                     ],
-                    [str(3), 0, 0, 0, 0.314003748427, 0, 0.998791145574],
+                    [str(3), 0, 0, 0, 0.31400374842, 0, 0.998791145574],
                 ]
 
                 if i == 0:
@@ -245,10 +246,10 @@ class TestFitDihedralWithGomc(BaseTest):
                     ],
                     [
                         str("0_3"),
-                        0.314003748427,
+                        0.31400374842,
                         0,
                         0,
-                        -0.157001874214,
+                        -0.15700187421,
                         0,
                         0,
                         0,
@@ -306,10 +307,10 @@ class TestFitDihedralWithGomc(BaseTest):
                     ],
                     [
                         str("0_3"),
-                        0.157001874212,
-                        0.471005622638,
+                        0.15700187421,
+                        0.47100562263,
                         0,
-                        -0.62800749685,
+                        -0.62800749684,
                         0,
                         0,
                         0.998791145574,
@@ -335,7 +336,7 @@ class TestFitDihedralWithGomc(BaseTest):
                             ) == mdf_math.round_to_sig_figs(
                                 correct_line_values[i][j], number_sig_i
                             )
-
+    
     def test_gaussian_log_file_fit_oplsaa_fit_ethane_HC_CT_CT_HC_with_2_log_files(
         self,
     ):
@@ -362,6 +363,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=1e-03,
+            opls_force_k0_zero=True
         )
 
         assert (
@@ -503,7 +505,7 @@ class TestFitDihedralWithGomc(BaseTest):
                         "k4_kcal_per_mol",
                         "r_squared",
                     ],
-                    [str(3), 0, 0, 0, 0.314003748425, 0, 0.998791145574],
+                    [str(3), 0, 0, 0, 0.31400374842, 0, 0.998791145574],
                 ]
 
                 if i == 0:
@@ -557,10 +559,10 @@ class TestFitDihedralWithGomc(BaseTest):
                     ],
                     [
                         str("0_3"),
-                        0.314003748425,
+                        0.31400374842,
                         0,
                         0,
-                        -0.157001874212,
+                        -0.15700187421,
                         0,
                         0,
                         0,
@@ -618,10 +620,10 @@ class TestFitDihedralWithGomc(BaseTest):
                     ],
                     [
                         str("0_3"),
-                        0.157001874212,
-                        0.471005622638,
+                        0.15700187421,
+                        0.47100562263,
                         0,
-                        -0.62800749685,
+                        -0.62800749684,
                         0,
                         0,
                         0.998791145574,
@@ -647,7 +649,7 @@ class TestFitDihedralWithGomc(BaseTest):
[…]
