+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -50,7 +50,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=1e-03,
-            opls_force_k0_zero=True
+            opls_force_k0_zero=True,
         )
 
         assert (
@@ -336,7 +336,7 @@ class TestFitDihedralWithGomc(BaseTest):
                             ) == mdf_math.round_to_sig_figs(
                                 correct_line_values[i][j], number_sig_i
                             )
-    
+
     def test_gaussian_log_file_fit_oplsaa_fit_ethane_HC_CT_CT_HC_with_2_log_files(
         self,
     ):
@@ -363,7 +363,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=1e-03,
-            opls_force_k0_zero=True
+            opls_force_k0_zero=True,
         )
 
         assert (
@@ -649,7 +649,7 @@ class TestFitDihedralWithGomc(BaseTest):
                             ) == mdf_math.round_to_sig_figs(
                                 correct_line_values[i][j], number_sig_i
                             )
-    
+
     def test_gaussian_log_file_fit_oplsaa_protonated_fragment_CT_CT_C_OH_in_COOH(
         self,
     ):
@@ -673,7 +673,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=1e-03,
-            opls_force_k0_zero=True
+            opls_force_k0_zero=True,
         )
 
         assert (
@@ -857,7 +857,7 @@ class TestFitDihedralWithGomc(BaseTest):
                         0,
                         0.925047594422,
                         1.32819621359,
-                        0.290365264694, 
+                        0.290365264694,
                         0,
                         0.998573206243,
                     ],
@@ -941,7 +941,7 @@ class TestFitDihedralWithGomc(BaseTest):
                         180.0,
                         0,
                         180.0,
-                        -0.371866278201 ,
+                        -0.371866278201,
                     ],
                     [
                         str("0_2"),
@@ -1139,7 +1139,7 @@ class TestFitDihedralWithGomc(BaseTest):
                         180.0,
                         0,
                         180.0,
-                        0.999295535579 ,
+                        0.999295535579,
                     ],
                 ]
 
@@ -1301,7 +1301,7 @@ class TestFitDihedralWithGomc(BaseTest):
                             ) == mdf_math.round_to_sig_figs(
                                 correct_line_values[i][j], number_sig_i
                             )
-    
+
     def test_gaussian_style_files_fit_oplsaa_fit_CT_CT_C_OH_in_COOH_missing_1st_point(
         self,
     ):
@@ -1326,7 +1326,7 @@ class TestFitDihedralWithGomc(BaseTest):
             gomc_cpu_cores=1,
             r_squared_min=0.99,
             r_squared_rtol=0.02,
-            opls_force_k0_zero=True
+            opls_force_k0_zero=True,
         )
 
         assert (
@@ -1969,7 +1969,7 @@ class TestFitDihedralWithGomc(BaseTest):
                                 atol=0.02,
                                 rtol=0.08,
                             )
-    
+
[…]

[…]
