### 2024-08-03 a1a82de
fixed path to GOMC .. to the container
