+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -17,7 +17,7 @@ gomc_binary_directory = "/opt/gomc_build/GOMC/bin"
 
 # User changable variable, as it needs to be run locally
 # Examples
-gomc_binary_directory = "/home/brad/Programs/GOMC/GOMC-2.75a/bin"
+# gomc_binary_directory = "/home/brad/Programs/GOMC/GOMC-2.75a/bin"
 # gomc_binary_directory = "/Users/calcraven/Documents/Vanderbilt/Research/MoSDeF/Dihedral_Fitter/GOMC/bin"
 
 
