### 2024-08-03 7f2562c
changed docs to say k0=constant or 0 in /math_operations.py
