+++ mosdef_dihedral_fit/utils/math_operations.py
@@ -424,7 +424,7 @@ def opls_dihedral(cos_powers_phi_and_constants_data, k0, k1, k2, k3, k4):
 
             A list of the OPLS k0 data, which is always 0 or 1 in this case.
             If all 0 values --> (0, 0, .., 0), then k0=0.
-            If all 1 values --> (1, 1, .., 1), then 0=constant
+            If all 1 values --> (1, 1, .., 1), then k0=constant
 
             # It is critical that 'const_1_minus_Cos_0_phi' be all (0, 0, .., 0) for k0=0
             # and all (1, 1, .., 1) 'const_1_minus_Cos_0_phi' for k0=constant
@@ -458,9 +458,10 @@ def opls_dihedral(cos_powers_phi_and_constants_data, k0, k1, k2, k3, k4):
 
     k0: int, or float, in energy units (i.e., kcal/mol, kJ/mol, Kelvin, ...)
         The 'k0' value is the k-value for the opls dihedral where n=0,
-        or the constant without a cosine multiple.
-        NOTE: In this case, it is set to zero (0) regardless of the
-        user entered value, because it is not in the equation form.
+        or the constant without a cosine multiple. The 'k0' can be a 
+        constant or zero (0). 
+        Please see 'all_sum_opls_const_1_plus_or_minus_cos_n_list' variable 
+        for more details.
     k1: int, or float, in energy units (i.e., kcal/mol, kJ/mol, Kelvin, ...)
         The 'k1' value is the k-value for the opls dihedral where n=1
         in the cosine multiple.
@@ -638,16 +639,6 @@ def opls_dihedral(cos_powers_phi_and_constants_data, k0, k1, k2, k3, k4):
 
     # check for acceptable cos_powers
     if cos_powers_modified not in [
-        "1",
-        "2",
-        "3",
-        "4",
-        "1_3",
-        "2_4",
-        "1_2",
-        "3_4",
-        "1_2_3",
-        "1_2_3_4",
         "1",
         "2",
         "3",
@@ -1032,8 +1023,8 @@ def opls_dihedral_n_1_2_3_4(phi_data, k_0, k_1, k_2, k_3, k_4):
         The 'phi_data' angle of the dihedral is in degrees
     k_0: int, or float, in energy units (i.e., kcal/mol, kJ/mol, Kelvin, ...)
         The 'k0' value is the k-value for the opls dihedral where n=0,
-        or the constant without a cosine multiple.
-        NOTE: In this case, it is set to zero (0) regardless of the
+        or the constant without a cosine multiple. The 'k0' can be a 
+        constant or zero (0). 
         user entered value, because it is not in the equation form.
     k_1: int, or float, in energy units (i.e., kcal/mol, kJ/mol, Kelvin, ...)
         The 'k1' value is the k-value for the opls dihedral where n=1
