### 2024-08-03 32bc6e6
[pre-commit.ci] auto fixes from pre-commit.com hooks for more information, see https://pre-commit.ci
