+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -15,7 +15,7 @@ from mosdef_dihedral_fit.tests.base_test import BaseTest
 # automated GitHub test binary path.
 # The automated GitHub test filename and location = "MoSDeF-dihedral-fit/.github/workflows/CI.yml"
 gomc_binary_directory = (
-    None  # set GOMC binary path here if you want to manually specify it
+    ""  # set GOMC binary path here if you want to manually specify it
 )
 # User changable variable, if the binary is not in the environment variable "PATH"
 # Examples
