### 2024-08-04 8ab3c74
Update test_fit_dihedral_with_gomc.py
Removed () from “” the file that sets the GOMC binary path here
