### 2024-08-04 87532f2
Update test_fit_dihedral_with_gomc.py
Removed “GOMC/bin” as I thing the file is in the path
