### 2024-08-04 f7a94f9
Update test_fit_dihedral_with_gomc.py
Added name of binary file= “GOMC_CPU_NVT" to look for in testing in GOMC installed
