### 2024-08-04 0bcb050
Update mosdef_dihedral_fit/dihedral_fit/fit_dihedral_with_gomc.py
fixed typo

Co-authored-by: Co Quach <43968221+daico007@users.noreply.github.com>
