### 2024-08-04 1f5fb1f
Update test_fit_dihedral_with_gomc.py
updated "opls_force_k0_zero" type error output to match input

"ERROR: Please enter the 'opls_force_k0_zero' as a bool."
