### 2024-08-05 80b84c6
Implement new method for installing GOMC into conda environment
