### 2024-08-05 67fdd1a
add contributing guideline to docs
