### 2024-08-05 0bedf64
Add instruction to CI.yml
