### 2024-08-05 efacbc4
change order of actions so /Users/calcraven/miniconda3/envs/dihedral_fit311 exists
