+++ .github/workflows/CI.yml
@@ -43,9 +43,6 @@ jobs:
       - name: change permission on GOMC make file and run it
         run: cd /opt/gomc_build/GOMC/ && sudo chmod u+x metamake.sh  && sudo ./metamake.sh NVT
 
-      - name: Add GOMC path as symlink to conda bin
-        run: sudo ln -s /opt/GOMC_CPU_NVT $CONDA_PREFIX/bin
-
       - name: Checkout Branch / Pull Request
         uses: actions/checkout@v3
 
@@ -58,6 +55,9 @@ jobs:
 
       - name: Install Package
         run: python -m pip install -e .
+
+      - name: Add GOMC path as symlink to conda bin
+        run: sudo ln -s /opt/GOMC_CPU_NVT $CONDA_PREFIX/bin
 
       - name: Test (OS -> ${{ matrix.os }} / Python -> ${{ matrix.python-version }})
         run: python -m pytest -v --cov=mosdef_dihedral_fit --cov-report=xml --cov-append --cov-config=setup.cfg --color yes --pyargs mosdef_dihedral_fit
