### 2024-08-05 edcdf1c
Fix GOMC build symlink path
