### 2024-08-05 9433ec6
add notes about using conda or mamba for install
