### 2024-08-06 679c8c3
match contributing in docs with contributing in landing page
