### 2024-08-18 6f2a6c3
added the rest of the example tests and additional tests to catch errors.  Updated the docs text and removed errors and warnings when building the docs.
