+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -19,7 +19,7 @@ from mosdef_dihedral_fit.tests.base_test import BaseTest
 gomc_binary_directory = ""
 # User changable variable, if the binary is not in the environment variable "PATH"
 # Examples
-# gomc_binary_directory = "/home/brad/Programs/GOMC/GOMC-2.75a/bin"
+gomc_binary_directory = "/home/brad/Programs/GOMC/GOMC-2.75a/bin"
 # gomc_binary_directory = "/Users/calcraven/Documents/Vanderbilt/Research/MoSDeF/Dihedral_Fitter/GOMC/bin"
 if gomc_binary_directory == "":
     gomc_binary_directory = subprocess.check_output(
@@ -661,7 +661,7 @@ class TestFitDihedralWithGomc(BaseTest):
                             ) == mdf_math.round_to_sig_figs(
                                 correct_line_values[i][j], number_sig_i
                             )
-
+    
     def test_gaussian_log_file_fit_oplsaa_protonated_fragment_CT_CT_C_OH_in_COOH(
         self,
     ):
@@ -5025,7 +5025,7 @@ class TestFitDihedralWithGomc(BaseTest):
                 r_squared_rtol=2,
                 opls_force_k0_zero=True,
             )
-
+    
     def test_error_r_squared_min_and_r_squared_rtol_need_adjusted(
         self,
     ):
@@ -5042,16 +5042,16 @@ class TestFitDihedralWithGomc(BaseTest):
             f"Abs\(delta\) = absolute_value\(R-squared_fitted - R-squared_new_dihedral\) \n"
             f"- The fits for all are shown here for all the dihedral combinations \n"
             f"\[opls_constants_used, R-squared_fitted, R-squared_new_dihedral, Abs\(delta\)\] \n"
-            f"\[\['1', -0.27129793, -1.2824201, 1.0111222\], "
-            f"\['2', 0.45284933, -0.069821698, 0.52267103\], "
-            f"\['3', -1.5045949, -3.102543, 1.5979481\], "
-            f"\['4', -1.5448789, -3.1964352, 1.6515563\], "
-            f"\['1_3', -0.0060455503, -0.85802999, 0.85198444\], "
-            f"\['2_4', 0.55469005, 0.092830036, 0.46186001\], "
-            f"\['3_4', -0.8973394, -2.3260485, 1.4287091\], "
-            f"\['1_2', 0.98670161, 0.96338608, 0.02331553\], "
-            f"\['1_2_3', 0.99785712, 0.98674454, 0.01111258\], "
-            f"\['1_2_3_4', 0.99807697, 0.98725523, 0.01082174\]\], \n"
+            f"\[\['1', -0.26874241, -1.2725105, 1.0037681\], "
+            f"\['2', 0.44842113, -0.07712682, 0.52554795\], "
+            f"\['3', -1.5086367, -3.0946117, 1.585975\], "
+            f"\['4', -1.5454298, -3.1867649, 1.6413351\], "
+            f"\['1_3', -0.0037937181, -0.85110572, 0.847312\], "
+            f"\['2_4', 0.55194616, 0.088277735, 0.46366842\], "
+            f"\['3_4', -0.89796747, -2.3178875, 1.41992\], "
+            f"\['1_2', 0.98674311, 0.96364339, 0.02309972\], "
+            f"\['1_2_3', 0.99792638, 0.98698695, 0.01093943\], "
+            f"\['1_2_3_4', 0.99810376, 0.98668897, 0.01141479\]\], \n"
             f"The 'r_squared_min' and 'r_squared_rtol' "
             f"variables may need to be adjusted, \n"
             f"there is likely something wrong with the fitting procedure, the "
@@ -5085,7 +5085,7 @@ class TestFitDihedralWithGomc(BaseTest):
[…]
