+++ docs/topic_guides/data_structures.rst
@@ -4,14 +4,18 @@
 Data Structures
 ===============
 
+
 Calculate a Dihedral Angle
 --------------------------
 
-.. automodule:: mosdef_dihedral_fit.utils.math_operations
-        :members: dihedral_angle
+	.. automodule:: mosdef_dihedral_fit.utils.math_operations
+        	:members: dihedral_angle
 
 
 Fit a Dihedral
 --------------
 
-	.. autofunction:: mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc.fit_dihedral_with_gomc
+	.. automodule:: mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc
+		:members: fit_dihedral_with_gomc
+
+
