+++ docs/getting_started/quick_start/fit_a_dihedral_with_Gaussian_and_write_all_files.rst
@@ -5,11 +5,34 @@ Fit the dihedral for Ethane using a Gaussian log file
 -----------------------------------------------------
 The ethane HC-CT-CT-HC dihedral fit example is provided below, where nine (9) of the HC-CT-CT-HC dihedral are fit simultaneously.
 
-
 .. note::
     The GOMC software need to be installed manually, outside of this Python install,
     with it's directory/path specified in the dihedral fit function.
 
+.. note::
+    **'atom_type_naming_style'** variable [str, optional, default= **'all_unique'**, (**'general'** or **'all_unique'**)].
+
+    Using **'general'** is preferred as it makes it easier to create the force field (FF) and is much 
+    more human readable.  For CHARMM-style FFs, like GOMC and NAMD, this means that all the general 
+    atom/bead types that are in an atom class must have the same non-bonded VDW parameters 
+    (i.e., for LJ for sigma and epsilon), but are allowed to have different partial charges; 
+    otherwise, it will cause errors. This is because the partial charges are listed in the PSF file 
+    for each atom individually, and not in the NAMD or GOMC FF file (.inp). 
+    Typically, the bonded parameters are all handled as using general atom/bead classes 
+    (Example: the carbon atoms in butane (C in CH3 and C in CH2) would have the same VDW and bonded 
+    parameters, but will have different partial charges). The 'general' setting allows the fitting 
+    of a group of atom/bead types together as an same atom class, where atom/bead class is 
+    determined in the FF XML file for each atom/bead type.
+    
+    Using **'all_unique'** basically sets each atom type as its own separate atom class.  Therefore, 
+    there is no distinction between atom type and atom class. This dihedral fit will on take into 
+    account the specific atom type, and not fit the general class (ignoring any FF XML file settings), 
+    which may make some dihedral fits more difficult. However, this could be the desired outcome in 
+    some situations.  Note: The MoSDeF force field does allow specific atom type dihedrals and other 
+    bonded parameters, which override the atom class values, but all the atoms in the bonded FF 
+    parameters must be the atom types. 
+    
+    atom_type_naming_style = 'general'
 
 Perform the dihedral fit using the Gaussian log file as the Quantum Mechanics (QM) engine
 and GOMC Molecular Mechanics (MM) engine, and write out all the viable dihedral fits via the standard
@@ -21,7 +44,7 @@ Import the required packages.
 .. code:: ipython3
 
     import unyt as u
-    from mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc import fit_dihedral_with_gomc
+    #from mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc import fit_dihedral_with_gomc
     import os
 
 Select the desired variables, file, and set the temperature.
@@ -42,8 +65,14 @@ Select the desired variables, file, and set the temperature.
     # selected automatically; however, you can override it if you are unsure of the setting.
     combining_rule = 'geometric'
 
-    # Atom type naming convention ( str, optional, default=’all_unique’, (‘general’ or ‘all_unique’) )
-    # General is safe and recommended since we are using a single FF XML file.
+    # Atom type naming convention [str, optional, default='all_unique', ('general' or 'all_unique')].
+    # SEE TOP OF PAGE FOR MORE DETAILS.
+    # -----------------------------------------------------------------------------------------------
+    # The 'general' setting allows the fitting 
+    # of a group of atom/bead types together as an same atom class, where atom/bead class is 
+    # determined in the FF XML file for each atom/bead type.
+    # -----------------------------------------------------------------------------------------------
+    # Using 'all_unique' basically sets each atom type as its own separate atom class.  
     atom_type_naming_style = 'general'
 
     # The GOMC binary path.
[…]
