+++ docs/topic_guides/data_structures.rst
@@ -17,5 +17,3 @@ Fit a Dihedral
 
 	.. automodule:: mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc
 		:members: fit_dihedral_with_gomc
-
-
