### 2024-08-18 0cff19f
Update test_fit_dihedral_with_gomc.py
