+++ mosdef_dihedral_fit/tests/test_math_operation.py
@@ -11,6 +11,8 @@ class TestMathOperations(BaseTest):
     # ********************************************
 
     def test_round_to_sig_figs(self):
+        assert mdf_math.round_to_sig_figs(0) == 0
+
         test_sigfig_input_value = 122.283638245
         expected_values = [
             0.0,
@@ -161,12 +163,12 @@ class TestMathOperations(BaseTest):
 
     def test_angle_between_2_vectors_errors(self):
         # Error 1
-        with pytest.raises(
-            ValueError,
-            match=f"ERROR: The 'vector_1' or 'vector_2', and  |vector_1||vector_2| == 0, which means the \n"
+        error_message = (
+            f"ERROR: The 'vector_1' or 'vector_2', and  |vector_1||vector_2| == 0, which means the \n"
             f"angle can not be calculated due to the divisor being zero in the formula; \n"
-            f"angle = arccos\[\(vector_1 dot vector_2\)/\(|vector_1||vector_2|\)\]",
-        ):
+            f"angle = arccos\[\(vector_1 dot vector_2\)/\(|vector_1||vector_2|\)\]"
+        )
+        with pytest.raises(ValueError, match=error_message):
             test_vector_1_input_value = [0, 0, 0]
             test_vector_2_input_value = [1, 1, 1]
             test_return_value = mdf_math.angle_between_2_vectors(
@@ -174,12 +176,7 @@ class TestMathOperations(BaseTest):
             )
 
         # Error 2
-        with pytest.raises(
-            ValueError,
-            match=f"ERROR: The 'vector_1' or 'vector_2', and  |vector_1||vector_2| == 0, which means the \n"
-            f"angle can not be calculated due to the divisor being zero in the formula; \n"
-            f"angle = arccos\[\(vector_1 dot vector_2\)/\(|vector_1||vector_2|\)\]",
-        ):
+        with pytest.raises(ValueError, match=error_message):
             test_vector_1_input_value = [1, 1, 1]
             test_vector_2_input_value = [0, 0, 0]
             test_return_value = mdf_math.angle_between_2_vectors(
@@ -456,3 +453,10 @@ class TestMathOperations(BaseTest):
     # ********************************************
     # test the dihedral_angle (END)
     # ********************************************
+
+    def test_check_previous_qu_values_match():
+        # Without error
+        all_value_list = [1, 2, 3, 4, 5]
+        current_value = [1, 2, 3, 4, 5]
+
+        # With error
