+++ mosdef_dihedral_fit/tests/test_file_read_write.py
@@ -1,6 +1,7 @@
 import os
 
 import numpy as np
+import pytest
 import unyt as u
 
 import mosdef_dihedral_fit.utils.math_operations as mdf_math
@@ -62,6 +63,15 @@ class TestFileReading(BaseTest):
             "H",
             "H",
         ]
+
+        error_fn = "with_errors/CT_CT_C_3_OH_with_errors.mol2"
+        full_error_fn = self.get_fn(error_fn)
+        error_msg = (
+            "ERROR: The provided mol2 format is not the required VMD TRIPOS format "
+            "or the mol2 file has zero \(0\) atoms or beads in it."
+        )
+        with pytest.raises(TypeError, match=error_msg):
+            get_atom_names_and_elements_from_mol2(full_error_fn)
 
     def test_get_from_pdb(self):
         fn = "gaussian_style_output_files/CT_CT_C_OH/input/starting_coords/CT_CT_C_3_OH.pdb"
