### 2024-08-19 c9c7808
fix math operation test
