### 2024-08-19 dea1acd
add exception test for pdb writer
