### 2024-08-19 GOMC-WSU/MoSDeF-dihedral-fit#97 @daico007
Added in #103
