### 2024-08-19 f61a550
Testing exceptions (#106)
* remove unused imports

* fix typo

* try remove io.py

* add some extra test

* fix math operation test

* add test mol2 files

* add matching error msg

* add exception test for pdb writer
