### 2024-08-19 GOMC-WSU/MoSDeF-dihedral-fit#99 @daico007
Addressed in #108
