### 2024-08-23 eb0fe44
added all the functions, but with warnings, saying input tests correctness is not checked.
