+++ mosdef_dihedral_fit/utils/math_operations.py
@@ -1,5 +1,4 @@
 import math
-
 import numpy as np
 
 
@@ -38,7 +37,6 @@ def round_to_sig_figs(value, sig_figs=3):
 # dihedral angles calculations
 # (START)
 # **************************************************************
-
 
 def normalize_vector(vector):
     """Generates the normalized vector from a given vector.
@@ -123,7 +121,7 @@ def dihedral_angle(
     atom_xyz_coord_1,
     atom_xyz_coord_2,
     atom_xyz_coord_3,
-    atom_xyz_coord_4,
+    atom_xyz_coord_4
 ):
     """Gets the dihedral angle between the four (4) atom coordinates given in cartesian coordinates.
 
@@ -149,6 +147,47 @@ def dihedral_angle(
     dihedral_angle_degrees: float (in degrees)
         The dihedral angle, in degrees, between the four (4) atoms.
     """
+    # check if any atom coordinates are 3 ints or floats in lists 
+    if  (
+        ( 
+            not isinstance(atom_xyz_coord_1, list)  
+            or not isinstance(atom_xyz_coord_2, list) 
+            or not isinstance(atom_xyz_coord_3, list)
+            or not isinstance(atom_xyz_coord_4, list)
+        ) 
+        or
+        (
+            len(atom_xyz_coord_1) != 3
+            or len(atom_xyz_coord_2) != 3
+            or len(atom_xyz_coord_3) != 3
+            or len(atom_xyz_coord_4) != 3
+        )
+        or 
+        ( 
+            not isinstance(atom_xyz_coord_1[0], (int, float))
+            or not isinstance(atom_xyz_coord_1[1], (int, float))
+            or not isinstance(atom_xyz_coord_1[2], (int, float))
+            or not isinstance(atom_xyz_coord_2[0], (int, float))
+            or not isinstance(atom_xyz_coord_2[1], (int, float))
+            or not isinstance(atom_xyz_coord_2[2], (int, float))
+            or not isinstance(atom_xyz_coord_3[0], (int, float))
+            or not isinstance(atom_xyz_coord_3[1], (int, float))
+            or not isinstance(atom_xyz_coord_3[2], (int, float))
+            or not isinstance(atom_xyz_coord_4[0], (int, float))
+            or not isinstance(atom_xyz_coord_4[1], (int, float))
+            or not isinstance(atom_xyz_coord_4[2], (int, float))
+        )
+    ):
+        raise ValueError(
+            f"ERROR: The one or more of the atom coordinates are not in the form of a "
+            f"list with three (3) ints or floats. \n"
+            f"Example: [1.2, 3.4, 5.6] \n "
+            f"atom_xyz_coord_1 = {atom_xyz_coord_1} and type = {type(atom_xyz_coord_1)}; \n"
+            f"atom_xyz_coord_2 = {atom_xyz_coord_2} and type = {type(atom_xyz_coord_2)}; \n"
+            f"atom_xyz_coord_3 = {atom_xyz_coord_3} and type = {type(atom_xyz_coord_3)}; \n"
+            f"atom_xyz_coord_4 = {atom_xyz_coord_4} and type = {type(atom_xyz_coord_4)}. "
+        )
+
     # check if any atom coordinates are the same
     if (
         list(atom_xyz_coord_1) == list(atom_xyz_coord_2)
@@ -220,7 +259,6 @@ def dihedral_angle(
 # (End)
 # **************************************************************
 
-
 def check_previous_qm_values_match(
     all_value_list, current_value, value_name, qm_engine, log_file_name
 ):
[…]
