+++ mosdef_dihedral_fit/tests/test_math_operation.py
@@ -9,7 +9,6 @@ class TestMathOperations(BaseTest):
     # ********************************************
     # test the mdf_math.round_to_sig_figs (START)
     # ********************************************
-
     def test_round_to_sig_figs(self):
         assert mdf_math.round_to_sig_figs(0) == 0
 
@@ -450,6 +449,407 @@ class TestMathOperations(BaseTest):
                 test_vectors_input_value[3],
             )
 
+    def test_dihedral_angle_error_coord_1_not_list(self):
+        # Test case 1
+        with pytest.raises(
+            ValueError,
+            match=f"ERROR: The one or more of the atom coordinates are not in the form of a "
+            f"list with three \(3\) ints or floats. \n"
+            f"Example: \[1.2, 3.4, 5.6\] \n "
+            f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'str'>; \n"
+            f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+        ):
+            test_vectors_input_value = [
+                '[1, 2, 3]',
+                [4, 5, 6],
+                [7, 8, 9],
+                [10, 11, 12],
+            ]
+            test_return_value = mdf_math.dihedral_angle(
+                test_vectors_input_value[0],
+                test_vectors_input_value[1],
+                test_vectors_input_value[2],
+                test_vectors_input_value[3],
+            )
+    
+    def test_dihedral_angle_error_coord_2_not_list(self):
+        # Test case 1
+        with pytest.raises(
+            ValueError,
+            match=f"ERROR: The one or more of the atom coordinates are not in the form of a "
+            f"list with three \(3\) ints or floats. \n"
+            f"Example: \[1.2, 3.4, 5.6\] \n "
+            f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'str'>; \n"
+            f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+        ):
+            test_vectors_input_value = [
+                [1, 2, 3],
+                '[4, 5, 6]',
+                [7, 8, 9],
+                [10, 11, 12],
+            ]
+            test_return_value = mdf_math.dihedral_angle(
+                test_vectors_input_value[0],
+                test_vectors_input_value[1],
+                test_vectors_input_value[2],
+                test_vectors_input_value[3],
+            )
+
+    def test_dihedral_angle_error_coord_3_not_list(self):
+        # Test case 1
+        with pytest.raises(
+            ValueError,
+            match=f"ERROR: The one or more of the atom coordinates are not in the form of a "
+            f"list with three \(3\) ints or floats. \n"
+            f"Example: \[1.2, 3.4, 5.6\] \n "
+            f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'list'>; \n"
+            f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'str'>; \n"
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+        ):
+            test_vectors_input_value = [
+                [1, 2, 3],
+                [4, 5, 6],
[…]
