+++ docs/topic_guides/data_structures.rst
@@ -10,10 +10,13 @@ Calculate a Dihedral Angle
 
 	.. automodule:: mosdef_dihedral_fit.utils.math_operations
         	:members: dihedral_angle
-
+			
 
 Fit a Dihedral
 --------------
 
 	.. automodule:: mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc
 		:members: fit_dihedral_with_gomc
+
+		
+

[…]
