+++ mosdef_dihedral_fit/utils/math_operations.py
@@ -1,4 +1,5 @@
 import math
+
 import numpy as np
 
 
@@ -37,6 +38,7 @@ def round_to_sig_figs(value, sig_figs=3):
 # dihedral angles calculations
 # (START)
 # **************************************************************
+
 
 def normalize_vector(vector):
     """Generates the normalized vector from a given vector.
@@ -118,10 +120,7 @@ def angle_between_2_vectors(vector_1, vector_2):
 
 
 def dihedral_angle(
-    atom_xyz_coord_1,
-    atom_xyz_coord_2,
-    atom_xyz_coord_3,
-    atom_xyz_coord_4
+    atom_xyz_coord_1, atom_xyz_coord_2, atom_xyz_coord_3, atom_xyz_coord_4
 ):
     """Gets the dihedral angle between the four (4) atom coordinates given in cartesian coordinates.
 
@@ -147,23 +146,21 @@ def dihedral_angle(
     dihedral_angle_degrees: float (in degrees)
         The dihedral angle, in degrees, between the four (4) atoms.
     """
-    # check if any atom coordinates are 3 ints or floats in lists 
-    if  (
-        ( 
-            not isinstance(atom_xyz_coord_1, list)  
-            or not isinstance(atom_xyz_coord_2, list) 
+    # check if any atom coordinates are 3 ints or floats in lists
+    if (
+        (
+            not isinstance(atom_xyz_coord_1, list)
+            or not isinstance(atom_xyz_coord_2, list)
             or not isinstance(atom_xyz_coord_3, list)
             or not isinstance(atom_xyz_coord_4, list)
-        ) 
-        or
-        (
+        )
+        or (
             len(atom_xyz_coord_1) != 3
             or len(atom_xyz_coord_2) != 3
             or len(atom_xyz_coord_3) != 3
             or len(atom_xyz_coord_4) != 3
         )
-        or 
-        ( 
+        or (
             not isinstance(atom_xyz_coord_1[0], (int, float))
             or not isinstance(atom_xyz_coord_1[1], (int, float))
             or not isinstance(atom_xyz_coord_1[2], (int, float))
@@ -259,6 +256,7 @@ def dihedral_angle(
 # (End)
 # **************************************************************
 
+
 def check_previous_qm_values_match(
     all_value_list, current_value, value_name, qm_engine, log_file_name
 ):
@@ -328,7 +326,7 @@ def sum_opls_const_1_plus_or_minus_cos_n_values(phi_list):
     | const_1_minus_Cos_4_phi = sum of the list using all phis [(1 - cos(4 * phi))] in k4 * (1 - cos(4 * phi))
 
     .. note::
-        These values are used to simplify the diheral fitting process, 
+        These values are used to simplify the diheral fitting process,
         allowing the user to deal with four (4) OPLS constants (see below).
 
 
@@ -338,7 +336,7 @@ def sum_opls_const_1_plus_or_minus_cos_n_values(phi_list):
         U_{opls-dihedral} = 1/2 * (
         k0 + k1 * [1 + cos[1 * phi]
         + k2 * [1 - cos[2 * phi]
-    
+
         + k3 * [1 + cos[3 * phi]
         + k4 * [1 - cos[4 * phi]
         )
@@ -349,7 +347,7 @@ def sum_opls_const_1_plus_or_minus_cos_n_values(phi_list):
[…]
