+++ mosdef_dihedral_fit/tests/test_math_operation.py
@@ -459,10 +459,10 @@ class TestMathOperations(BaseTest):
             f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'str'>; \n"
             f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'list'>; \n"
-            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. ",
         ):
             test_vectors_input_value = [
-                '[1, 2, 3]',
+                "[1, 2, 3]",
                 [4, 5, 6],
                 [7, 8, 9],
                 [10, 11, 12],
@@ -473,7 +473,7 @@ class TestMathOperations(BaseTest):
                 test_vectors_input_value[2],
                 test_vectors_input_value[3],
             )
-    
+
     def test_dihedral_angle_error_coord_2_not_list(self):
         # Test case 1
         with pytest.raises(
@@ -484,11 +484,11 @@ class TestMathOperations(BaseTest):
             f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'str'>; \n"
             f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'list'>; \n"
-            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. ",
         ):
             test_vectors_input_value = [
                 [1, 2, 3],
-                '[4, 5, 6]',
+                "[4, 5, 6]",
                 [7, 8, 9],
                 [10, 11, 12],
             ]
@@ -509,12 +509,12 @@ class TestMathOperations(BaseTest):
             f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'str'>; \n"
-            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. "
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'list'>. ",
         ):
             test_vectors_input_value = [
                 [1, 2, 3],
                 [4, 5, 6],
-                '[7, 8, 9]',
+                "[7, 8, 9]",
                 [10, 11, 12],
             ]
             test_return_value = mdf_math.dihedral_angle(
@@ -534,13 +534,13 @@ class TestMathOperations(BaseTest):
             f"atom_xyz_coord_1 = \[1, 2, 3\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_2 = \[4, 5, 6\] and type = <class 'list'>; \n"
             f"atom_xyz_coord_3 = \[7, 8, 9\] and type = <class 'list'>; \n"
-            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'str'>. "
+            f"atom_xyz_coord_4 = \[10, 11, 12\] and type = <class 'str'>. ",
         ):
             test_vectors_input_value = [
                 [1, 2, 3],
                 [4, 5, 6],
                 [7, 8, 9],
-                '[10, 11, 12]',
+                "[10, 11, 12]",
             ]
[…]
