### 2024-08-23 dc03e4d
added all the functions, but with warnings, saying input tests correctness is not checked.
