### 2024-08-23 c55002e
added all the functions, but with warnings, saying input tests correctness is not checked.
