### 2024-08-23 e669c16
added all the functions, but with warnings, saying input tests correctness is not checked.
