### 2024-08-23 d5ca0a1
fixed citations on JOSS paper
