+++ paper/paper.bib
@@ -83,7 +83,7 @@ journal = {J. Comp. Chem.},
 volume = {7},
 issue = {2},
 pages = {230-252},
-year = {986},
+year = {1986},
 doi = {10.1002/jcc.540070216},
 URL = {https://doi.org/10.1002/jcc.540070216},
 }
@@ -121,40 +121,6 @@ year = {1903},
 volume = {11},
 pages = {657},
 doi = {10.1002/andp.19033160802}
-}
-
-# CHARMM FF part 1
-@article{Brooks:2009,
-author = {Brooks, B. R. and Brooks, C. L. and Mackerell, A. D. and Nilsson, L. and Petrella, R. J. and Roux, B. and Won, Y. and Archontis, G. and Bartels, C. and Boresch, S. and Caflisch, A. and Caves, L. and Cui, Q. and Dinner, A. R. and Feig, M. and Fischer, S. and Gao, J. and Hodoscek, M. and Im, W. and Kuczera, K. and Lazaridis, T. and Ma, J. and Ovchinnikov, V. and Paci, E. and Pastor, R. W. and Post, C. B. and Pu, J. Z. and Schaefer, M. and Tidor, B. and Venable, R. M. and Woodcock, H. L. and Wu, X. and Yang, W. and York, D. M. and Karplus, M.},
-doi = {10.1002/jcc.21287},
-issn = {1096987X},
-journal = {J.~Comput.\ Chem.},
-keywords = {Biomolecular simulation,Biophysical computation,CHARMM program,Energy function,Molecular dynamics,Molecular mechanics,Molecular modeling},
-month = {jul},
-number = {10},
-pages = {1545--1614},
-pmid = {19444816},
-publisher = {John Wiley and Sons Inc.},
-title = {{CHARMM: The biomolecular simulation program}},
-volume = {30},
-year = {2009}
-}
-
-# CHARMM FF part 2
-@article{Lee:2016-CG,
-author = {Lee, Jumin and Cheng, Xi and Swails, Jason M. and Yeom, Min Sun and Eastman, Peter K. and Lemkul, Justin A. and Wei, Shuai and Buckner, Joshua and Jeong, Jong Cheol and Qi, Yifei and Jo, Sunhwan and Pande, Vijay S. and Case, David A. and Brooks, Charles L. and MacKerell, Alexander D. and Klauda, Jeffery B. and Im, Wonpil},
-doi = {10.1021/acs.jctc.5b00935},
-issn = {15499626},
-journal = {J. Chem. Theor. Comput.},
-month = {jan},
-number = {1},
-pages = {405--413},
-pmid = {26631602},
-publisher = {American Chemical Society},
-title = {{CHARMM-GUI Input Generator for NAMD, GROMACS, AMBER, OpenMM, and CHARMM/OpenMM Simulations Using the CHARMM36 Additive Force Field}},
-url = {https://pubs.acs.org/doi/full/10.1021/acs.jctc.5b00935},
-volume = {12},
-year = {2016}
 }
 
 # Mixing rule geometric and arithmetic epslion
@@ -209,7 +175,7 @@ url = {https://github.com/mosdef-hub/gmso},
 
 # MoSDeF forcefield-utilities
 @fidgit{forcefield-utilities:2022,
-author =   {},
+author =   {Umesh Timalsina},
 title = {forcefield-utilities},
 year = {2022},
 publisher = "Github",
@@ -385,8 +351,8 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
 
 # vmd-python
 @fidgit{vmd-python:2016,
-author = {},
-title = {vmd-python},
+author = {Robin Betz},
+title = {VMD-python},
 year = {2016},
 publisher = {Github},
 url = {https://github.com/Eigenstate/vmd-python},
