### 2024-08-24 ba8d289
added zeroing a dihedral example and citations.
