+++ paper/paper.bib
@@ -112,6 +112,19 @@ pages = {11225--11236},
 doi = {10.1021/ja9621760}
 }
 
+# OPLS4 FF
+@article{Chao:2021,
+author = {Lu, Chao and Wu, Chuanjie and Ghoreishi, Delaram and Chen, Wei and Wang, Lingle and Damm, Wolfgang and Ross, Gregory A. and Dahlgren, Markus K. and Russell, Ellery and Von Bargen, Christopher D. and Abel, Robert and Friesner, Richard A. and Harder, Edward D.},
+title = {OPLS4: Improving Force Field Accuracy on Challenging Regimes of Chemical Space},
+journal = {Journal of Chemical Theory and Computation},
+volume = {17},
+number = {7},
+pages = {4291-4300},
+year = {2021},
+doi = {10.1021/acs.jctc.1c00302},
+URL = {https://doi.org/10.1021/acs.jctc.1c00302},
+}
+
 # Mie FF
 @article{Mie:1903,
 author = {Mie, G.},
@@ -349,11 +362,24 @@ URL = {https://doi.org/10.1021/acs.jpcb.5b01354},
   annote={This symposium took place between March 19-21, 2001. {ISBN} 1-58113-292-1.}
 }
 
-# vmd-python
+# vmd-python GitHub
 @fidgit{vmd-python:2016,
 author = {Robin Betz},
 title = {VMD-python},
 year = {2016},
 publisher = {Github},
 url = {https://github.com/Eigenstate/vmd-python},
+}
+
+# COOH dihedral with zeroing - Potoff
+@article{Ganesh:2004,
+author = {Kamath, Ganesh and Cao, Feng and Potoff, Jeffrey J.},
+title = {An Improved Force Field for the Prediction of the Vapor−Liquid Equilibria for Carboxylic Acids},
+journal = {The Journal of Physical Chemistry B},
+volume = {108},
+number = {37},
+pages = {14130-14136},
+year = {2004},
+doi = {10.1021/jp048581s},
+URL = {https://doi.org/10.1021/jp048581s},
 }
