### 2024-08-24 73bddb9
adds other ways to get FF or dihedral parameters
