+++ paper/paper.bib
@@ -383,3 +383,69 @@ year = {2004},
 doi = {10.1021/jp048581s},
 URL = {https://doi.org/10.1021/jp048581s},
 }
+
+# Other dihedral fit methods and difficulties with fits
+@article{Kania:2021,
+author = {Kania, Adrian and Sarapata, Krzysztof and Gucwa, Michał and Wójcik-Augustyn, Anna},
+title = {Optimal Solution to the Torsional Coefficient Fitting Problem in Force Field Parametrization},
+journal = {The Journal of Physical Chemistry A},
+volume = {125},
+number = {12},
+pages = {2673-2681},
+year = {2021},
+doi = {10.1021/acs.jpca.0c10845},
+URL = {https://doi.org/10.1021/acs.jpca.0c10845}
+
+}
+
+# Other ML dihedral fit methods 
+@article{Friederich:2018,
+author = {Friederich, Pascal and Konrad, Manuel and Strunk, Timo and Wenzel, Wolfgang},
+title = {Machine learning of correlated dihedral potentials for atomistic molecular force fields},
+journal = {Scientific Reports},
+volume = {8},
+issue = {1},
+pages = {2559},
+year = {2018},
+doi = {10.1038/s41598-018-21070-0},
+URL = {https://doi.org/10.1038/s41598-018-21070-0}
+}
+
+# Other ML vibrational fit dihedral 
+@article{Vermeyen:2023,
+author = {Vermeyen, Tom and Cunha, Ana and Bultinck, Patrick and Herrebout, Wouter},
+title = {Impact of conformation and intramolecular interactions on vibrational circular dichroism spectra identified with machine learning},
+journal = {Communications Chemistry},
+volume = {6},
+issue = {1},
+pages = {148},
+year = {2023},
+doi = {10.1038/s42004-023-00944-z},
+URL = {https://doi.org/10.1038/s42004-023-00944-z}
+}
+
+# CHARMM FFTK using vibrational spectra
+@article{Mayne:2013,
+author = {Christopher G. Mayne and Jan Saam and Klaus Schulten and Emad Tajkhorshid and James C. Gumbart},
+title = {Rapid parameterization of small molecules using the Force Field Toolkit},
+journal = {Journal of Computational Chemistry},
+volume = {34},
+issue = {32},
+pages = {2757-2770},
+year = {2013},
+doi = {10.1002/jcc.23422},
+URL = {https://doi.org/10.1002/jcc.23422}
+}
+
+# dihedral - bonded parameters of GROMOS
+@article{Schmid:2011,
+author = {Schmid, Nathan and Eichenberger, Andreas P. and Choutko, Alexandra and Riniker, Sereina and Winger, Moritz and Mark, Alan E. and van Gunsteren, Wilfred F.},
+title = {Definition and testing of the GROMOS force-field versions 54A7 and 54B7},
+journal = {European Biophysics Journal},
+volume = {40},
+issue = {7},
+pages = {843},
+year = {2011},
+doi = {10.1007/s00249-011-0700-9},
+URL = {https://doi.org/10.1007/s00249-011-0700-9}
+}
