+++ paper/paper.bib
@@ -398,7 +398,7 @@ URL = {https://doi.org/10.1021/acs.jpca.0c10845}
 
 }
 
-# Other ML dihedral fit methods 
+# Other ML dihedral fit methods
 @article{Friederich:2018,
 author = {Friederich, Pascal and Konrad, Manuel and Strunk, Timo and Wenzel, Wolfgang},
 title = {Machine learning of correlated dihedral potentials for atomistic molecular force fields},
@@ -411,7 +411,7 @@ doi = {10.1038/s41598-018-21070-0},
 URL = {https://doi.org/10.1038/s41598-018-21070-0}
 }
 
-# Other ML vibrational fit dihedral 
+# Other ML vibrational fit dihedral
 @article{Vermeyen:2023,
 author = {Vermeyen, Tom and Cunha, Ana and Bultinck, Patrick and Herrebout, Wouter},
 title = {Impact of conformation and intramolecular interactions on vibrational circular dichroism spectra identified with machine learning},
