### 2024-08-24 2234d4f
adds other ways to get FF or dihedral parameters
