### 2024-08-24 58e0211
added cites for all of 1st paragraph, and for 1-4 scaling and non-bonded combining rules
