+++ paper/paper.bib
@@ -428,7 +428,7 @@ URL = {https://doi.org/10.1038/s42004-023-00944-z}
 @article{Mayne:2013,
 author = {Christopher G. Mayne and Jan Saam and Klaus Schulten and Emad Tajkhorshid and James C. Gumbart},
 title = {Rapid parameterization of small molecules using the Force Field Toolkit},
-journal = {Journal of Computational Chemistry},
+journal = {J Comput Chem.},
 volume = {34},
 issue = {32},
 pages = {2757-2770},
@@ -448,4 +448,210 @@ pages = {843},
 year = {2011},
 doi = {10.1007/s00249-011-0700-9},
 URL = {https://doi.org/10.1007/s00249-011-0700-9}
+}
+
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# cause issues in dihedral transferablity
+@article{Huang:2013,
+author = {Huang, Lei and Roux, Benoît},
+title = {Automated Force Field Parameterization for Nonpolarizable and Polarizable Atomic Models Based on Ab Initio Target Data},
+journal = {Journal of Chemical Theory and Computation},
+volume = {9},
+number = {8},
+pages = {3543-3556},
+year = {2013},
+doi = {10.1021/ct4003477},
+URL = {https://doi.org/10.1021/ct4003477}
+}
+
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# cause issues in dihedral transferablity
+@article{Mackerell:2004,
+author = {Alexander D. Mackerell Jr.},
+title = {"Empirical force fields for biological macromolecules: Overview and issues"},
+journal = {J Comput Chem.},
+volume = {25},
+issue = {13},
+pages = {1584-1604},
+year = {2004},
+doi = {10.1002/jcc.20082},
+URL = {https://doi.org/10.1002/jcc.20082}
+}
+
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# cause issues in dihedral transferablity
+@article{Vanommeslaeghe:2010,
+author = {K. Vanommeslaeghe and E. Hatcher and C. Acharya and S. Kundu and S. Zhong and J. Shim and E. Darian and O. Guvench and P. Lopes and I. Vorobyov and A. D. MacKerell, Jr.},
+title = {CHARMM general force field: A force field for drug-like molecules compatible with the CHARMM all-atom additive biological force fields},
+journal = {J Comput Chem.},
+volume = {31},
+issue = {4},
+pages = {671-690},
+year = {2010},
+doi = {10.1002/jcc.21367},
+URL = {https://doi.org/10.1002/jcc.21367}
+}
+
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# cause issues in dihedral transferablity
+# and dihedral fitting methods
+@article{Vanommeslaeghe:2014,
+author = {Kenno Vanommeslaeghe and Olgun Guvench and Alexander D. MacKerell, Jr.},
+title = {Molecular Mechanics},
+journal = {Current Pharmaceutical Design},
+volume = {20},
+issue = {20},
+pages = {3281-3292},
+year = {2014},
+doi = {10.2174/13816128113199990600},
+}
+
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges) and 1-n non-bonded interactions
+# cause issues in dihedral transferablity
+@article{Chen:2015,
+author = {Siyan Chen and Shasha Yi and Wenmei Gao and Chuncheng Zuo and Zhonghan Hu},
+title = {Force field development for organic molecules: Modifying dihedral and 1-n pair interaction parameters},
[…]
