+++ paper/paper.bib
@@ -450,7 +450,7 @@ doi = {10.1007/s00249-011-0700-9},
 URL = {https://doi.org/10.1007/s00249-011-0700-9}
 }
 
-# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges)
 # cause issues in dihedral transferablity
 @article{Huang:2013,
 author = {Huang, Lei and Roux, Benoît},
@@ -464,7 +464,7 @@ doi = {10.1021/ct4003477},
 URL = {https://doi.org/10.1021/ct4003477}
 }
 
-# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges)
 # cause issues in dihedral transferablity
 @article{Mackerell:2004,
 author = {Alexander D. Mackerell Jr.},
@@ -478,7 +478,7 @@ doi = {10.1002/jcc.20082},
 URL = {https://doi.org/10.1002/jcc.20082}
 }
 
-# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges)
 # cause issues in dihedral transferablity
 @article{Vanommeslaeghe:2010,
 author = {K. Vanommeslaeghe and E. Hatcher and C. Acharya and S. Kundu and S. Zhong and J. Shim and E. Darian and O. Guvench and P. Lopes and I. Vorobyov and A. D. MacKerell, Jr.},
@@ -492,7 +492,7 @@ doi = {10.1002/jcc.21367},
 URL = {https://doi.org/10.1002/jcc.21367}
 }
 
-# FF's 1-4 (non-bonded) VDW and electrostatic (charges) 
+# FF's 1-4 (non-bonded) VDW and electrostatic (charges)
 # cause issues in dihedral transferablity
 # and dihedral fitting methods
 @article{Vanommeslaeghe:2014,
