### 2024-08-24 0d065c1
renamed 'other_functions' to 'additional_utilities' in the docs'
