### 2024-08-25 c5a64a7
Made original lines 12-14 and 15 in the paper clearer to the reader as suggested
