+++ paper/paper.bib
@@ -656,7 +656,7 @@ doi = {10.1038/s41563-021-01015-1},
 URL = {https://doi.org/10.1038/s41563-021-01015-1},
 }
 
-# Molecular simulations books 
+# Molecular simulations books
 @Inbook{Allen:2017,
 author={Allen, M. P. and Tildesley, D. J},
 title="Computer simulation of liquids (2nd ed.)",
@@ -664,11 +664,10 @@ year={2017},
 publisher={Oxford University Press},
 }
 
-# Molecular simulations books 
+# Molecular simulations books
 @Inbook{Frekel:2002,
 author={Daan Frenkel and Berend Smit},
 title={Understanding Molecular Simulation From Algorithms to Applications (2nd ed.)},
 year={2002},
 publisher={Academic Press},
 }
-
