+++ paper/paper.bib
@@ -659,7 +659,7 @@ URL = {https://doi.org/10.1038/s41563-021-01015-1},
 # Molecular simulations books
 @Inbook{Allen:2017,
 author={Allen, M. P. and Tildesley, D. J},
-title="Computer simulation of liquids (2nd ed.)",
+title="Computer Simulation of Liquids (2nd ed.)",
 year={2017},
 publisher={Oxford University Press},
 }
