### 2024-08-25 2606cce
Made original lines 12-14 and 15 in the paper clearer to the reader as suggested
