### 2024-08-25 481330e
reorganized and removed duplicate statements
