### 2024-08-25 7087171
reorganized and removed duplicate statements
