### 2024-08-25 d29c2ff
reorganized and removed duplicate statements and fixed some docs typos
