### 2024-08-25 c932964
add more read/write tests
