### 2024-08-25 09ebc81
errors on get_gaussian_log_file_data
