### 2024-08-26 94a5956
add remainder of the tests
