+++ mosdef_dihedral_fit/tests/test_file_read_write.py
@@ -336,6 +336,19 @@ class TestFileReading(BaseTest):
         write_qm_data_files({full_path: out_indices})
         assert os.path.exists("extracted_gaussian_data/dihedral.txt")
 
+        # qm_engine not in list
+        with pytest.raises(ValueError):
+            write_qm_data_files(None, qm_engine="None")
+
+        # when qm_engine="gaussian_style_final_files"
+        # manual_dihedral_atom_numbers_list must be a list of len 4
+        with pytest.raises(TypeError):
+            write_qm_data_files(
+                {full_path: out_indices},
+                None,
+                qm_engine="gaussian_style_final_files",
+            )
+
     def test_get_matching_dihedral_info_and_opls_fitting_data(self):
         out = get_matching_dihedral_info_and_opls_fitting_data(
             fit_dihedral_atom_types=["HC", "CT", "CT", "HC"],
@@ -414,6 +427,102 @@ class TestFileReading(BaseTest):
         ]
         assert np.allclose(opls_paramsList[0], expected_opls_params)
 
+        # fit_dihedral_atom_types must be a list of length 4
+        with pytest.raises(TypeError):
+            out = get_matching_dihedral_info_and_opls_fitting_data(
+                fit_dihedral_atom_types=["HC", "CT", "CT"],
+                psf_path_and_filename=self.get_fn(
+                    "gaussian/HC_CT_CT_HC/output/GOMC_pdb_psf_ff_files.psf"
+                ),
+                qm_log_file_dict={
+                    self.get_fn(
+                        "gaussian/HC_CT_CT_HC/output/HC_CT_CT_HC_multiplicity_1.log"
+                    ): []
+                },
+                mol2_file=self.get_fn(
+                    "gaussian/HC_CT_CT_HC/input/starting_coords/ethane_aa.mol2"
+                ),
+                qm_engine="gaussian",
+                manual_dihedral_atom_numbers_list=None,
+            )
+
+        # Each element of fit_dihedral_atom_types must be a string
+        with pytest.raises(TypeError):
+            out = get_matching_dihedral_info_and_opls_fitting_data(
+                fit_dihedral_atom_types=[1, 2, 3, 4],
+                psf_path_and_filename=self.get_fn(
+                    "gaussian/HC_CT_CT_HC/output/GOMC_pdb_psf_ff_files.psf"
+                ),
+                qm_log_file_dict={
+                    self.get_fn(
+                        "gaussian/HC_CT_CT_HC/output/HC_CT_CT_HC_multiplicity_1.log"
+                    ): []
+                },
+                mol2_file=self.get_fn(
+                    "gaussian/HC_CT_CT_HC/input/starting_coords/ethane_aa.mol2"
+                ),
+                qm_engine="gaussian",
+                manual_dihedral_atom_numbers_list=None,
+            )
+
+        # fit_dihedral_atom_types must be a list
+        with pytest.raises(TypeError):
+            out = get_matching_dihedral_info_and_opls_fitting_data(
+                fit_dihedral_atom_types=None,
+                psf_path_and_filename=self.get_fn(
+                    "gaussian/HC_CT_CT_HC/output/GOMC_pdb_psf_ff_files.psf"
+                ),
+                qm_log_file_dict={
+                    self.get_fn(
+                        "gaussian/HC_CT_CT_HC/output/HC_CT_CT_HC_multiplicity_1.log"
+                    ): []
+                },
+                mol2_file=self.get_fn(
[…]
