### 2024-08-26 GOMC-WSU/MoSDeF-dihedral-fit#100 @bc118
@RiesBen Hello!  I hope you are doing well.  We are working to finish this,  have made a lot of progress, and are getting close to completing it.  I do have the following question (s):

**_For this comment "Documentation: Repository Doc strings - API Reference Docs missing? (can be generated with sphinx-apidoc)":_**

- The user's  main functions existed but are **_not labeled as functions but  [data structures](https://mosdef-dihedral-fit.readthedocs.io/en/latest/topic_guides/data_structures.html)_** for the software, and they cover all the possible combinations of variables that the user can enter,  checking them and providing very informative error outputs.  The main functions are (i.e., data structures) .

- I added all the **_[additional utilities](https://mosdef-dihedral-fit.readthedocs.io/en/latest/topic_guides/additional_utilities.html)_**, which are not intended to be used for the direct application of this software.  Is this what you were suggesting here or did you only want the main user function (i.e., data structures in the above)?

It would be much appreciated if you could clarify this so I can make sure it is what you are suggesting  Can you or let me know if adding the _**additional utilities**_ what you were thinking or simply keeping the _**data structure**_s  (i.e., main functions)?

Thank you again for your time on your review and your comments that made this software package better!
