### 2024-08-26 GOMC-WSU/MoSDeF-dihedral-fit#100 @RiesBen
Hi @bc118 ,
a right, this is the section I was searching for! cool, that you added the additional utilities! :)
I think both in something like API reference or so would be the correct section, no?
Why is it actually Datastructure? Isn't this more a code structure or API reference section?
