### 2024-08-26 GOMC-WSU/MoSDeF-dihedral-fit#100 @bc118
Hi!  Ah, yes.  In the docs titles/labeling, we have tried to keep it consistent across the MoSDeF software family (mbuild, foyer, gmso, forcefield-utilities, and MoSDeF-GOMC, etc.) and Data structures was what we have been using for this section.  I was not around for the original naming of the sections, but I suspect it means how they need to structure the input data and what to expect the output or attributes in the classes.

I also added a few other features (force to k0=0 for OPLS form or a constant) and a lot more examples.
