### 2024-08-27 GOMC-WSU/MoSDeF-dihedral-fit#98 @daico007
Hi @naik-aakash, I have added more tests specifically on the exceptions raised in our main code in #106 and #120. Please let us know if you think if there's any area that we missed and needed more testing.
