### 2024-08-27 GOMC-WSU/MoSDeF-dihedral-fit#98 @bc118
I will add that some of the missing coverage parts in code that I write come from “catch all” checks; in which case, I do not know how the code would end up in that error or warning.  However, if someone makes an error in the code, something that changes in in unexpected ways or allows the user to enter problematic options, it may catch some of these conditions.
