### 2024-08-31 a400f6b
bug fix for other forms of Gaussian 16 dihedral output files
