### 2024-08-31 b4ca95a
bug fix for other forms of Gaussian 16 dihedral output files
