+++ mosdef_dihedral_fit/tests/test_file_read_write.py
@@ -185,6 +185,41 @@ class TestFileReading(BaseTest):
         assert elements == ["C", "C", "H", "H", "H", "H", "H", "H"]
         assert n_atoms == 8
 
+        # Wrong input types
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(None, [3, 1, 2, 8])
+
+        # Dict value must be list of integers
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(
+                {full_path: [0.1, 0.2, 0.3, -0.4]}, [3, 1, 2, 8]
+            )
+
+        # Dict key must be a string
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(
+                {1: [0.1, 0.2, 0.3, 0.4]}, [3, 1, 2, 8]
+            )
+
+        # Dict value must be a list
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(
+                {full_path: "ERROR"}, [3, 1, 2, 8]
+            )
+
+        # manual_dihedral_atom_numbers_list must be a list of len 4
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(
+                {full_path: list(np.arange(32, dtype=int))}, None
+            )
+
+        # manual_dihedral_atom_numbers_list must be a list of integers
+        with pytest.raises(TypeError):
+            get_final_gaussian_output_file_data(
+                {full_path: list(np.arange(32, dtype=int))},
+                [0.1, 0.2, 0.3, 0.4],
+            )
+
     def test_get_gaussian_log_file_data_1(self):
         full_path = self.get_fn(
             "gaussian/CT_CT_C_OH/output/CT_CT_C_OH_multiplicity_1.log"
