### 2024-09-02 7ef02ab
added some absolute tolerance instead of relative tolerance.  Added R**2 = -1 for any that are negative as simplier and then atol to 2 , so you can override the R**2 error if you want not also.
