+++ mosdef_dihedral_fit/dihedral_fit/fit_dihedral_with_gomc.py
@@ -37,7 +37,7 @@ def fit_dihedral_with_gomc(
     atom_type_naming_style="general",
     gomc_cpu_cores=1,
     r_squared_min=0.98,
-    r_squared_rtol=2.5e-02,
+    r_squared_atol=2.5e-02,
     opls_force_k0_zero=False,
 ):
     """Fit the desired dihedral to a MM force field, based on QM data.
@@ -319,14 +319,15 @@ def fit_dihedral_with_gomc(
         dihedral energies are set to zero.
 
         NOTE: This value may need adjusted to get the dihedral fit to solve correctly.
-    r_squared_rtol: float (0 < r_squared_min < 1), default=2.5e-02
-        Where the QM data is defined as the actual data; this is the difference
+    r_squared_atol: float (0 < r_squared_min < 2), default=2.5e-02
+        Where the QM data is defined as the actual data; this is the absolute difference
         of the dihedral's calculated R-squared values between:
         * The QM-MM fitting process, where the fit MM dihedral k-values are zero (0).
         * The MM calculations where the fit k-value are entered in the MM data and
         compared to the QM data.
 
         NOTE: This value may need adjusted to get the dihedral fit to solve correctly.
+        NOTE: The max is 2 as the min R**2 is -1, if the fit it is worse than the data average.
     opls_force_k0_zero: bool, default=False
         The k0 constant is from the equation listed above.
         If True, this is force sets the k0 constant in the opls equation to zero (0),
@@ -696,17 +697,17 @@ def fit_dihedral_with_gomc(
             f"but it must be a 0<float<1."
         )
 
-    # test the 'r_squared_rtol' input
-    if isinstance(r_squared_rtol, float):
-        if not (r_squared_rtol > 0 and r_squared_rtol < 1):
+    # test the 'r_squared_atol' input
+    if isinstance(r_squared_atol, float):
+        if not (r_squared_atol > 0 and r_squared_atol < 2):
             raise ValueError(
-                f"ERROR: The 'r_squared_rtol' = {r_squared_rtol}, "
-                f"but it must be a 0<float<1."
+                f"ERROR: The 'r_squared_atol' = {r_squared_atol}, "
+                f"but it must be a 0<float<2."
             )
 
     else:
         raise TypeError(
-            f"ERROR: The 'r_squared_rtol' is a {type( r_squared_rtol)}, "
+            f"ERROR: The 'r_squared_atol' is a {type( r_squared_atol)}, "
             f"but it must be a 0<float<1."
         )
 
@@ -3345,13 +3346,17 @@ def fit_dihedral_with_gomc(
                 ),
             )
         ]
-
+        r_squared_k0_is_constant = mdf_math.get_r_squared(
+            sorted_Gaussian_minus_GOMC_data_total_energy_kcal_per_mol_normalized_list,
+            fit_opls_dihedral_k0_is_constant,
+        )
         # get R**2 for the fit, running through GOMC to get the new energy of the
         # individual fit.
         opls_r_squared_fitted_data_via_gomc_iter = mdf_math.get_r_squared(
-            gaussian_data_total_energy_kcal_per_mol_normalize_list,
             GOMC_data_fitted_total_energy_kcal_per_mol_normalize_list,
+            gaussian_data_total_energy_kcal_per_mol_normalize_list,
         )
+
         opls_r_squared_fitted_data_via_gomc_list.append(
             opls_r_squared_fitted_data_via_gomc_iter
         )
@@ -3392,7 +3397,7 @@ def fit_dihedral_with_gomc(
         # Compare original fit vs run through EACH GOMC as a validation test case
         if opls_q == 0:
[…]
