### 2024-09-02 7e9bebe
added a HC-CT-CT-HC test on butane to ensure all the duplicate angles are being calculated properly
