+++ mosdef_dihedral_fit/tests/test_fit_dihedral_with_gomc.py
@@ -7388,7 +7388,6 @@ class TestFitDihedralWithGomc(BaseTest):
                     for j in range(0, len(correct_line_values[i])):
                         # check the string listing what cosin powers are used
                         if j == 0:
-                            print
                             assert str(split_line_i[j]) == str(
                                 correct_line_values[i][j]
                             )
@@ -9802,6 +9801,319 @@ class TestFitDihedralWithGomc(BaseTest):
                         0.489877706752,
                         0,
                         0.999580885743,
+                    ],
+                ]
+
+                if i == 0:
+                    assert split_line_i == correct_line_values[i]
+
+                else:
+                    assert len(split_line_i) == len(correct_line_values[i])
+                    for j in range(0, len(correct_line_values[i])):
+                        # check the string listing what cosin powers are used
+                        if j == 0:
+                            assert str(split_line_i[j]) == str(
+                                correct_line_values[i][j]
+                            )
+
+                        # check the k-values and the r-squared fit
+                        else:
+                            assert np.isclose(
+                                mdf_math.round_to_sig_figs(
+                                    float(split_line_i[j]), number_sig_i
+                                ),
+                                mdf_math.round_to_sig_figs(
+                                    correct_line_values[i][j], number_sig_i
+                                ),
+                                atol=0.02,
+                                rtol=0.08,
+                            )
+
+    def test_gaussian_log_file_fit_butane_HC_CT_CT_HC_check_all_summed_same_angles_working(self):
+        fit_dihedral_with_gomc(
+            ["HC", "CT", "CT", "HC"],
+            self.get_fn(
+                "gaussian_style_output_files/HC_CT_CT_HC_butane/input/starting_coords/HC_CT_CT_HC_butane.mol2"
+            ),
+            self.get_fn("amber_aa_butane_CT_CT_CT_CT_gmso.xml"),
+            298.15 * u.Kelvin,
+            gomc_binary_directory,
+            {
+                self.get_fn(
+                    "gaussian_style_output_files/HC_CT_CT_HC_butane/output"
+                ): [],
+            },
+            manual_dihedral_atom_numbers_list=[6, 1, 2, 9],
+            qm_engine="gaussian_style_final_files",
+            combining_rule=None,
+            atom_type_naming_style="general",
+            gomc_cpu_cores=1,
+            r_squared_min=0.98,
+            r_squared_atol=0.02,
+            opls_force_k0_zero=False,
+        )
+
+        assert (
+            os.path.isfile("all_normalized_energies_in_kcal_per_mol.txt")
+            is True
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_1_2_3_4_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_1_2_3_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_1_2_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_1_3_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_1_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_2_4_in_kcal_per_mol.txt"
+            )
+            is False
+        )
+        assert (
+            os.path.isfile(
+                "all_normalized_energies_OPLS_fit_2_in_kcal_per_mol.txt"
[…]
