+++ .github/workflows/CI.yml
@@ -21,8 +21,16 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macOs-latest"]
         python-version: ["3.10", "3.11"]
+        os: ["ubuntu", "macos"]
+        include:
+          - os: ubuntu
+            environment-file: environment.yml
+            miniforge-variant: Mambaforge
+            miniforge-version: 4.9.2-4
+          - os: macos
+            environment-file: environment.yml
+            miniforge-variant: Mambaforge-pypy3
 
     defaults:
       run:
@@ -47,11 +55,12 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install Mamba
-        uses: mamba-org/setup-micromamba@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           environment-file: environment.yml
-          create-args: >-
-            python=${{ matrix.python-version }}
+          miniforge-variant: ${{ matrix.miniforge-variant }}
+          miniforge-version: ${{ matrix.miniforge-version }}
+          python-version: ${{ matrix.python-version }}
 
       - name: Install Package
         run: python -m pip install -e .
