+++ .github/workflows/CI.yml
@@ -27,7 +27,7 @@ jobs:
           - os: ubuntu
             environment-file: environment.yml
             miniforge-variant: Mambaforge
-            miniforge-version: 4.9.2-4
+            miniforge-version: 24.7.1-0
           - os: macos
             environment-file: environment.yml
             miniforge-variant: Mambaforge-pypy3
@@ -37,6 +37,16 @@ jobs:
         shell: bash -l {0}
 
     steps:
+      - name: Mamba build
+        - uses: actions/checkout@v2
+        - uses: conda-incubator/setup-miniconda@v2
+          with:
+            environment-file: ${{ matrix.environment-file }}
+            miniforge-variant: ${{ matrix.miniforge-variant }}
+            miniforge-version: ${{ matrix.miniforge-version }}
+            python-version: ${{ matrix.python-version }}
+            use-mamba: true
+
       - name: create dir
         run: sudo mkdir /opt/gomc_build
 
@@ -53,14 +63,6 @@ jobs:
 
       - name: Checkout Branch / Pull Request
         uses: actions/checkout@v3
-
-      - name: Install Mamba
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          environment-file: environment.yml
-          miniforge-variant: ${{ matrix.miniforge-variant }}
-          miniforge-version: ${{ matrix.miniforge-version }}
-          python-version: ${{ matrix.python-version }}
 
       - name: Install Package
         run: python -m pip install -e .
