+++ .github/workflows/CI.yml
@@ -1,4 +1,4 @@
-name: CI
+ame: CI
 
 # Controls when the workflow will run
 on:
@@ -21,32 +21,14 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
+        os: ["ubuntu-latest", "macOs-latest"]
         python-version: ["3.10", "3.11"]
-        os: ["ubuntu", "macos"]
-        include:
-          - os: ubuntu
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge
-            miniforge-version: 24.7.1-0
-          - os: macos
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge-pypy3
 
     defaults:
       run:
         shell: bash -l {0}
 
     steps:
-      - name: Install Mamba
-        - uses: actions/checkout@v2
-        - uses: conda-incubator/setup-miniconda@v2
-          with:
-            environment-file: ${{ matrix.environment-file }}
-            miniforge-variant: ${{ matrix.miniforge-variant }}
-            miniforge-version: ${{ matrix.miniforge-version }}
-            python-version: ${{ matrix.python-version }}
-            use-mamba: true
-
       - name: create dir
         run: sudo mkdir /opt/gomc_build
 
@@ -63,6 +45,18 @@ jobs:
 
       - name: Checkout Branch / Pull Request
         uses: actions/checkout@v3
+
+      - name: Install Mamba
+        uses: actions/checkout@v2
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          environment-file: ${{ matrix.environment-file }}
+          miniforge-variant: ${{ matrix.miniforge-variant }}
+          miniforge-version: ${{ matrix.miniforge-version }}
+          python-version: ${{ matrix.python-version }}
+          use-mamba: true
+          create-args: >-
+            python=${{ matrix.python-version }}
 
       - name: Install Package
         run: python -m pip install -e .
