### 2024-09-28 ad9e7f5
Update CI.yml
