+++ .github/workflows/CI.yml
@@ -1,4 +1,4 @@
-ame: CI
+name: CI
 
 # Controls when the workflow will run
 on:
