+++ .github/workflows/CI.yml
@@ -47,10 +47,10 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install Mamba
-        uses: actions/checkout@v2
-        uses: conda-incubator/setup-miniconda@v2
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
         with:
-          environment-file: ${{ matrix.environment-file }}
+          environment-file: ${{ matrix.environment }}
           miniforge-variant: ${{ matrix.miniforge-variant }}
           miniforge-version: ${{ matrix.miniforge-version }}
           python-version: ${{ matrix.python-version }}
