### 2024-09-28 e25a730
Update CI.yml
