+++ .github/workflows/CI.yml
@@ -21,8 +21,16 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macOs-latest"]
         python-version: ["3.10", "3.11"]
+        os: ["ubuntu", "macos"]
+        include:
+          - os: ubuntu
+            environment-file: environment.yml
+            miniforge-variant: Mambaforge
+            miniforge-version: 4.9.2-4
+          - os: macos
+            environment-file: environment.yml
+            miniforge-variant: Mambaforge-pypy3
 
     defaults:
       run:
