### 2024-09-28 ed49100
Update CI.yml
