+++ .github/workflows/CI.yml
@@ -31,12 +31,25 @@ jobs:
           - os: macos
             environment-file: environment.yml
             miniforge-variant: Mambaforge-pypy3
+    
 
     defaults:
       run:
         shell: bash -l {0}
 
     steps:
+      - name: Install Mamba
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          environment-file: ${{ matrix.environment }}
+          miniforge-variant: ${{ matrix.miniforge-variant }}
+          miniforge-version: ${{ matrix.miniforge-version }}
+          python-version: ${{ matrix.python-version }}
+          use-mamba: true
+          create-args: >-
+            python=${{ matrix.python-version }} 
+            
       - name: create dir
         run: sudo mkdir /opt/gomc_build
 
@@ -53,18 +66,6 @@ jobs:
 
       - name: Checkout Branch / Pull Request
         uses: actions/checkout@v3
-
-      - name: Install Mamba
-      - uses: actions/checkout@v2
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          environment-file: ${{ matrix.environment }}
-          miniforge-variant: ${{ matrix.miniforge-variant }}
-          miniforge-version: ${{ matrix.miniforge-version }}
-          python-version: ${{ matrix.python-version }}
-          use-mamba: true
-          create-args: >-
-            python=${{ matrix.python-version }}
 
       - name: Install Package
         run: python -m pip install -e .
