### 2024-09-28 e284919
Update CI.yml
