+++ .github/workflows/CI.yml
@@ -21,35 +21,14 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
+        os: ["ubuntu-latest", "macOs-latest"]
         python-version: ["3.10", "3.11"]
-        os: ["ubuntu", "macos"]
-        include:
-          - os: ubuntu
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge
-            miniforge-version: 4.9.2-4
-          - os: macos
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge-pypy3
-    
 
     defaults:
       run:
         shell: bash -l {0}
 
     steps:
-      - name: Install Mamba
-      - uses: actions/checkout@v2
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          environment-file: ${{ matrix.environment }}
-          miniforge-variant: ${{ matrix.miniforge-variant }}
-          miniforge-version: ${{ matrix.miniforge-version }}
-          python-version: ${{ matrix.python-version }}
-          use-mamba: true
-          create-args: >-
-            python=${{ matrix.python-version }} 
-            
       - name: create dir
         run: sudo mkdir /opt/gomc_build
 
@@ -66,6 +45,13 @@ jobs:
 
       - name: Checkout Branch / Pull Request
         uses: actions/checkout@v3
+
+      - name: Install Mamba
+        uses: mamba-org/setup-micromamba@v1
+        with:
+          environment-file: environment.yml
+          create-args: >-
+            python=${{ matrix.python-version }}
 
       - name: Install Package
         run: python -m pip install -e .
