+++ .github/workflows/CI.yml
@@ -103,3 +103,4 @@ jobs:
         with:
           push: true
           tags: ${{ env.DOCKER_TAGS }}
+          
