+++ .github/workflows/CI.yml
@@ -103,4 +103,3 @@ jobs:
         with:
           push: true
           tags: ${{ env.DOCKER_TAGS }}
-          
