### 2024-09-28 6706823
Update CI.yml
