+++ .github/workflows/CI.yml
@@ -47,7 +47,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install Mamba
-        uses: mamba-org/setup-micromamba@v1
+        uses: mamba-org/setup-micromamba@v1.9.0
         with:
           environment-file: environment.yml
           create-args: >-
