+++ .github/workflows/CI.yml
@@ -44,10 +44,10 @@ jobs:
         run: cd /opt/gomc_build/GOMC/ && sudo chmod u+x metamake.sh  && sudo ./metamake.sh NVT
 
       - name: Checkout Branch / Pull Request
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4.2.0
 
       - name: Install Mamba
-        uses: mamba-org/setup-micromamba@v1
+        uses: mamba-org/setup-micromamba@v1.9.0
         with:
           environment-file: environment.yml
           create-args: >-
