### 2024-09-28 e26c072
Update CI.yml
