+++ .github/workflows/CI.yml
@@ -44,7 +44,7 @@ jobs:
         run: cd /opt/gomc_build/GOMC/ && sudo chmod u+x metamake.sh  && sudo ./metamake.sh NVT
 
       - name: Checkout Branch / Pull Request
-        uses: actions/checkout@v4.2.0
+        uses: actions/checkout@v4
 
       - name: Install Mamba
         uses: mamba-org/setup-micromamba@v1.9.0
