+++ .github/workflows/CI.yml
@@ -47,6 +47,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Install Mamba
+        #uses: mamba-org/setup-micromamba@v1
         uses: conda-incubator/setup-miniconda@v3
         with:
           environment-file: environment.yml
