+++ .github/workflows/CI.yml
@@ -31,6 +31,7 @@ jobs:
           - os: macos
             environment-file: environment.yml
             miniforge-variant: Mambaforge-pypy3
+            
     defaults:
       run:
         shell: bash -l {0}
