### 2024-09-28 b7bfca2
Update CI.yml
