+++ .github/workflows/CI.yml
@@ -21,7 +21,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macOs-latest"]
+        os: ["ubuntu", "macos"]
         python-version: ["3.10", "3.11"]
         include:
           - os: ubuntu
