+++ .github/workflows/CI.yml
@@ -23,14 +23,7 @@ jobs:
       matrix:
         os: ["ubuntu", "macos"]
         python-version: ["3.10", "3.11"]
-        include:
-          - os: ubuntu
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge
-            miniforge-version: 4.9.2-4
-          - os: macos
-            environment-file: environment.yml
-            miniforge-variant: Mambaforge-pypy3
+
     defaults:
       run:
         shell: bash -l {0}
