+++ .github/workflows/CI.yml
@@ -21,7 +21,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macos-latest"]
+        os: ["ubuntu-latest", "macOs-latest"]
         python-version: ["3.10", "3.11"]
 
     defaults:
