### 2024-10-02 6eede6d
Update paper.md
Remove line 74, update line 72.
