### 2024-10-02 GOMC-WSU/MoSDeF-dihedral-fit#94 @jpotoff
We have decided to resolve this issue by deleting the paragraph at line number 74.  While this is a known phenomena in the field, once getting into the literature it was clear there isn't a neat way of doing this well within the limited space of the article.
