### 2024-10-03 59b9294
Update paper.md
Fixed typo
