### 2024-10-03 f8c2e6f
Update paper.md
Fixed typo
