### 2024-10-05 GOMC-WSU/MoSDeF-dihedral-fit#95 @bc118
We are trying to keep these consistent across the MoSDeF ecosystem (MoSDeF-dihedral-fit, MoSDeF-GOMC, GMSO, foyer, mbuild, etc.). Therefore, we will consider adding later when we revise all of them.
