### 2024-10-10 61c8b30
changed the wording on how to use the periodic form for CHARMM and that equations only go to 4th cos power/multiple.
