+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.0.3"
-release = "0.0.3"
+version = "0.1.0"
+release = "0.1.0"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
