+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.0.3
+current_version = 0.1.0
 commit = True
 tag = True
 message = Bump to version {new_version}
