### 2024-10-11 GOMC-WSU/MoSDeF-dihedral-fit#94 @bc118
@naik-aakash

A new mosdef-dihedral-fit version 0.1.0 was released on GitHub and conda-forge, which contains all the review requests. Additionally, new examples are listed on the README file and are posted [here](https://github.com/GOMC-WSU/GOMC_Examples/tree/main/MoSDeF-dihedral-fit)
