### 2024-10-12 GOMC-WSU/MoSDeF-dihedral-fit#94 @naik-aakash
Thanks @bc118 will try to have a look asap in next days 😄
