### 2024-10-12 GOMC-WSU/MoSDeF-dihedral-fit#94 @bc118
Ok,  thanks! 🙏
