### 2024-10-13 GOMC-WSU/MoSDeF-dihedral-fit#94 @bc118
@naik-aakash   Please be aware of this minor change which we would like to fix for the paper.  This was an oversight in the equally contributing tag in the authorship.  https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/131
