### 2024-10-15 #6970 @phibeck
Hi @jpotoff thank you for letting us know you've completed working through the reviewers' comments. :tada: I saw your change to the authorlist ([#131](https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/131)), which does not seem to be reflected in the latest proof. Could you cross-check that please? Thanks.

@RiesBen I see you've completed your checklist, thank you! :tada: Could you please confirm that you agree to the publication of this software in its current form?
@naik-aakash, could you please let me know if the changes are satisfactory, and if so, tick off the relevant items on your checklist?
