### 2024-10-27 #6970 @jpotoff
@phibeck Ok, working on the checklist.  We will attempt to update any missing DOIs, however, some of the materials cited (in many cases software) do not have DOIs, so those will have to remain as they are.
