### 2024-10-27 95abfe6
bug fix - randomly outputing misordered coord files because elements names not atom names used in the xyz files that were used to create the coor file.
