+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.1.1"
-release = "0.1.1"
+version = "0.1.2"
+release = "0.1.2"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
