### 2024-10-28 f07cc78
Need to lock GSD at gsd =3.3.2 until it is compatible (and others) with NumPy=>2.x
