+++ paper/paper.bib
@@ -197,6 +197,7 @@ title = {GMSO: General Molecular Simulation Object},
 year = {2019},
 publisher = {Github},
 url = {https://github.com/mosdef-hub/gmso},
+doi ={10.5281/zenodo.12533861},
 }
 
 # MoSDeF forcefield-utilities
@@ -206,6 +207,7 @@ title = {forcefield-utilities},
 year = {2022},
 publisher = "Github",
 url = {https://github.com/mosdef-hub/forcefield-utilities},
+doi ={10.5281/zenodo.10494867},
 }
 
 # MoSDeF part 1
