### 2024-11-08 db687f7
Update paper.bib
