### 2024-11-12 be5c06d
Update CITATION.cff
