### 2024-11-12 31dc9bb
Update CITATION.cff
