### 2024-11-12 2acf0bd
Update CITATION.cff
Add DOI to CITATION.cff
