+++ CITATION.cff
@@ -90,4 +90,4 @@ preferred-citation:
     orcid: 'https://orcid.org/0000-0002-4421-8787'
   title: "MoSDeF-dihedral-fit"
   year: "TBD"
-  doi: "TBD"
+  doi: "10.5281/zenodo.14107384"
