+++ paper/paper.bib
@@ -244,7 +244,9 @@ year = {2020}
 	title       = "MoSDeF-dihedral-fit: A simple software package to fit dihedrals via the MoSDeF software.",
 	year        = "2023",
 	publisher   = "Github",
-	url         = "https://github.com/GOMC-WSU/MoSDeF-dihedral-fit"
+	url         = {https://github.com/GOMC-WSU/MoSDeF-dihedral-fit},
+	doi = {10.5281/zenodo.14107384}
+
 }
 
 # MoSDeF-GOMC part 1
