### 2024-11-12 8920dac
Update paper.bib
Update DOI line 536
