+++ paper/paper.bib
@@ -533,7 +533,7 @@ volume = {36},
 issue = {6},
 pages = {376-384},
 year = {2015},
-doi = {doi.org/10.1002/jcc.23808},
+doi = {10.1002/jcc.23808},
 URL = {https://doi.org/10.1002/jcc.23808}
 }
 
