### 2024-11-12 #6970 @jpotoff
DOIs added for all items except for these:
No DOI given, and none found for title: Sur le mélange des gaz
- No DOI given, and none found for title: Gaussian 16 Revision C.01
- No DOI given, and none found for title: VMD-python
- No DOI given, and none found for title: Computer Simulation of Liquids (2nd ed.)
- No DOI given, and none found for title: Understanding Molecular Simulation From Algorithms...
