### 2024-11-12 #6970 @jpotoff
Software version name: MoSDeF-dihedral-fit 0.1.4
Software DOI: [10.5281/zenodo.14107384]
Confirmed author names, affiliations and ORCID
