### 2024-11-12 #6970 @jpotoff
License is consistent between software and Zenodo.
