+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.1.3"
-release = "0.1.3"
+version = "0.1.4"
+release = "0.1.4"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
