+++ mosdef_dihedral_fit/__init__.py
@@ -5,4 +5,4 @@ from mosdef_dihedral_fit.dihedral_fit.fit_dihedral_with_gomc import (
 )
 from mosdef_dihedral_fit.utils.math_operations import dihedral_angle
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
