### 2024-11-13 ca7eded
Bump to version 0.1.6
