### 2024-11-13 #6970 @phibeck
> Software version name: MoSDeF-dihedral-fit 0.1.4 Software DOI: [10.5281/zenodo.14107384] Confirmed author names, affiliations and ORCID

Hi @jpotoff I see there's a later release. Could you again confirm which version should be used for this paper?
