### 2024-11-13 #6970 @jpotoff
> > Software version name: MoSDeF-dihedral-fit 0.1.4 Software DOI: [10.5281/zenodo.14107384] Confirmed author names, affiliations and ORCID
>
> Hi @jpotoff I see there's a later release. Could you again confirm which version should be used for this paper?

Hi @phibeck we had a few hiccups in the Zenodo DOI generation process.  Those are now sorted that out and the correct version is 0.1.7 and the correct DOI is 10.5281/zenodo.14119720.
