### 2024-11-15 #6970 @phibeck
Hi @jpotoff could you please adjust the title of the archive to match that of the paper? I also noticed differences in affiliations between the paper and the archive. Could you make sure these match too, or else let me know if there's a reason these two are different? Thank you
