### 2024-11-15 #6970 @bc118
@phibeck Hello!  I hope all is well!  Happy Friday!

I can answer this because it was my previous suggestion.

We wanted these affiliations to match the other MoSDeF releases, standardizing them across MoSDeF eco system, which includes, but is not limited to:

- GMSO - https://zenodo.org/records/12533861
- mbuild - https://zenodo.org/records/7348318
- MoSDeF-GOMC - https://zenodo.org/records/7217632
-  forcefield-utilities - https://zenodo.org/records/10494867

The above zenodo repositories for the  MoSDeF eco system also have added affiliations in the corresponding journal publications, including the Vanderbilt university folks have multiple affiliations within Vanderbilt and its institutes.

On a secondary note, I believe that the CFF file only allows 1 affiliation per person from my previous search, unless you write it as a string.  If it is written as a string, it will not show the affiliate connections as well.  This is not the driving factor, but another consideration.

Thanks,

Brad
