### 2024-11-29 #6970 @phibeck
Hi @jpotoff, thanks for your patience. Here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.

- lines 6-11: in your affiliations, please spell out USA and UK
- line 38: "do not"
- line 41: "are not"
- line 42: "due to the different parameters"
- line 45: "allows users to quickly calculate"
- line 51: if I understand correctly, Gaussian16 .log is a file format, right? To distinguish file formats from regular text it can be useful to just mark them as `.log` I would probably write "provide a Gaussian16 `.log` file or Gaussian-like QM simulation file that covers..." XML can be either left as is or written as `.xml` too
- line 52: "(typically between 0 and 360 degrees)
- line 53: Again, I would recommend `.mol2
- line 56: remove hyphen in "(RB)-torsions" (and only introduce acronym if needed later, see line 98).
- line 63: MM acronym already defined above. Additionally, I would suggest "While many of these MM force field parameters such as bonds, angles and improper dihedrals (impropers) can be transferred between force fields, ..." I would also suggest to write "(often referred to as "impropers")" if that is correct.
- line 69: remove "absolutely"
- line 75: "there is a need for a"
- line 86: add an @ in front of "Jorgensen:1996" to render reference correctly
- lines 98f: I am not sure of the use of the equations in the paper. At present, they are neither referenced in the manuscript nor explained including all parameters, making them less helpful for the manuscript's intended audience (i.e., a diverse, non-specialist group).
- line 109: "1-n" with n in italic font
- line 116: capitalize Monte Carlo
- line 120: capitalize Monte Carlo
- line 126: capitalize Molecular Simulation Design Framework
- line 129: "n-alkane" with n in italic font
- line 182: check correct capitalization of title
- line 192: "n-alkanes" with n in italic font
