### 2024-11-30 d7dc2dd
added some suggestions or clarifications that the editor suggested.
