+++ paper/paper.bib
@@ -44,7 +44,7 @@ issn = {15206106},
 journal = {J. Phys. Chem. B},
 number = {14},
 pages = {2569--2577},
-title = {{Transferable potentials for phase equilibria. 1. United-atom description of n-alkanes}},
+title = {{Transferable potentials for phase equilibria. 1. United-atom description of \emph{n}-alkanes}},
 url = {https://pubs.acs.org/sharingguidelines},
 volume = {102},
 year = {1998}
@@ -174,7 +174,7 @@ year = {1970}
 # Mixing rule arithmetic sigma
 @article{Lorentz:1881,
 author = {Lorentz, H. A.},
-title = {Ueber die Anwendung des Satzes vom Virial in der kinetischen Theorie der Gase},
+title = {{U}eber die {A}nwendung des {S}atzes vom {V}irial in der kinetischen {T}heorie der {G}ase},
 journal = {Ann. d. Phys.},
 volume = {12},
 pages = {127--136},
@@ -213,7 +213,7 @@ doi = {10.5281/zenodo.10494867},
 # MoSDeF part 1
 @article{Cummings:2021,
 author = {Cummings, P.T. and McCabe, C. and Iacovella, C.R. and Ledeczi, A. and Jankowski, E. and Jayaraman, A. and Palmer, J.C. and Maginn, E.J. and Glotzer, S.C. and Anderson, J.A. and Siepmann, J.I. and  Potoff, J. and Matsumoto, R.A. and Gilmer, J.B. and DeFever, R.S. and Singh, R. and Crawford, B.},
-Title = {Open-Source Molecular Modeling Software in Chemical Engineering, with Focus on the Molecular Simulation Design Framework (MoSDeF)},
+Title = {Open-Source Molecular Modeling Software in Chemical Engineering, with Focus on the {M}olecular {S}imulation {D}esign {F}ramework ({M}o{SD}e{F})},
 journal = {AICHE J.},
 volume = {67(3)},
 pages = {e17206},
@@ -232,7 +232,7 @@ number = {3},
 pages = {1779--1793},
 pmid = {32004433},
 publisher = {American Chemical Society},
-title = {{MoSDeF, a Python Framework Enabling Large-Scale Computational Screening of Soft Matter: Application to Chemistry-Property Relationships in Lubricating Monolayer Films}},
+title = {{{M}o{SD}e{F}, a Python Framework Enabling Large-Scale Computational Screening of Soft Matter: Application to Chemistry-Property Relationships in Lubricating Monolayer Films}},
 url = {https://pubs.acs.org/doi/full/10.1021/acs.jctc.9b01183},
 volume = {16},
 year = {2020}
@@ -241,7 +241,7 @@ year = {2020}
 # MoSDeF-dihedral-fit GitHub
 @fidgit{Crawford:2023b,
 	author      = "Crawford, Brad and Quach, Co and Craven, Nicholas and Iacovella, Christopher R. and McCabe, Clare and Cummings, Peter T. and  Potoff, Jeffrey",
-	title       = "MoSDeF-dihedral-fit: A simple software package to fit dihedrals via the MoSDeF software.",
+	title       = "{M}o{SD}e{F}-dihedral-fit: A simple software package to fit dihedrals via the {M}o{SD}e{F} software.",
 	year        = "2023",
 	publisher   = "Github",
 	url         = {https://github.com/GOMC-WSU/MoSDeF-dihedral-fit},
@@ -252,7 +252,7 @@ year = {2020}
 # MoSDeF-GOMC part 1
 @article{Crawford:2023a,
[…]
