+++ paper/paper.md
@@ -40,36 +40,36 @@ authors:
     affiliation: 2
 
 affiliations:
- - name: Atomfold, PA, USA
+ - name: Atomfold, PA, United States of America
    index: 1
- - name: Department of Chemical Engineering and Materials Science, Wayne State University, Detroit, MI 48202-4050, USA
+ - name: Department of Chemical Engineering and Materials Science, Wayne State University, Detroit, MI 48202-4050, United States of America
    index: 2
- - name: Department of Chemical and Biomolecular Engineering, Vanderbilt University, Nashville, TN 37235-1604, USA
+ - name: Department of Chemical and Biomolecular Engineering, Vanderbilt University, Nashville, TN 37235-1604, United States of America
    index: 3
- - name: Multiscale Modeling and Simulation (MuMS) Center, Vanderbilt University, Nashville, TN 37212, USA
+ - name: Multiscale Modeling and Simulation (MuMS) Center, Vanderbilt University, Nashville, TN 37212, United States of America
    index: 4
- - name: Interdisciplinary Material Science Program, Vanderbilt University, Nashville, TN 37235-0106, USA
+ - name: Interdisciplinary Material Science Program, Vanderbilt University, Nashville, TN 37235-0106, United States of America
    index: 5
- - name: School of Engineering and Physical Sciences, Heriot-Watt University, Edinburgh EH14 4AS, UK
+ - name: School of Engineering and Physical Sciences, Heriot-Watt University, Edinburgh EH14 4AS, United Kingdom
    index: 6
 
-date: 10 October 2024
+date: 30 November 2024
 bibliography: paper.bib
 
 ---
 
 # Summary
 
-Molecular Mechanics (MM) simulations (e.g., molecular dynamics and Monte Carlo) provide a third method of scientific discovery, adding to the traditional theoretical and experimental scientific methods [@Mielke:2019; @Siegfried:2014].  Experimental methods measure the data under set conditions (e.g., temperature and pressure), whereas the traditional theoretical methods are based on analytical equations, and sometimes their constants are fitted to experimental data.  The MM simulations are deterministic and stochastic, and their models, commonly known as "force fields", can be optimized to match experimental data, similar to analytical theory-based methods [@Allen:2017; @Frekel:2002; @Jorgensen:1996; @Martin:1998; @Weiner:1984; @Weiner:1986; @Potoff:2009; @Hemmen:2015; @Errington:1999].  In larger, more complex systems, the stochastic simulation's molecules can jump large energy barriers that deterministic simulations may not be able to overcome in a reasonable timeframe, even with modern computing capabilities [@Allen:2017; @Frekel:2002].  However, deterministic and stochastic systems that provide adequate sampling for calculating a given property can provide critical insights into the system's phase space, which are not obtainable via traditional theoretical and experimental methods.  Additionally, molecular simulations provide critical insights from visualizations and by obtaining chemical or material properties that do not currently exist, are not easily attainable (e.g., too expensive or dangerous) by traditional theoretical and experimental methods [@Hollingsworth:2018; @Hirst:2014], or require hard-to-achieve conditions, such as very high pressures and temperatures [@Yu:2023; @Koneru:2022; @Swai:2020; @Kumar:2022; @Louie:2021].  However, the force field parameters are ideally determined from Quantum Mechanics (QM) simulations or other methods, including the vibrational spectrum and machine learning methods [@Kania:2021; @Friederich:2018; @Vermeyen:2023; @Mayne:2013; @Schmid:2011; @Vanommeslae
[…]
