### 2024-11-30 a9d6f6f
[pre-commit.ci] auto fixes from pre-commit.com hooks for more information, see https://pre-commit.ci
