+++ paper/paper.bib
@@ -281,7 +281,7 @@ journal = {SoftwareX},
 keywords = {Adsorption,GPU,Gibbs ensemble,Molecular simulation,Monte Carlo,Phase equilibrium},
 pages = {20--27},
 publisher = {Elsevier B.V.},
-title = {{GOMC: GPU Optimized {M}onte {C}arlo for the simulation of phase equilibria and physical properties of complex fluids}},
+title = {{{GOMC}: GPU Optimized {M}onte {C}arlo for the simulation of phase equilibria and physical properties of complex fluids}},
 url = {https://doi.org/10.1016/j.softx.2018.11.005},
 volume = {9},
 year = {2019}
