+++ paper/paper.bib
@@ -554,7 +554,6 @@ year="2019",
 publisher="Springer International Publishing",
 address="Cham",
 pages="39--47",
-abstract="Scientific research involves the formulation of theory to explain observed phenomena and using experimentation to test and evolve these theories. Over the past two decades, computational modeling and simulation (M{\&}S) has become accepted as the third leg of scientific research because it provides additional insights that often are impractical or impossible to acquire using theoretical and experimental analysis alone. The purpose of this chapter is to explore how M{\&}S is used in system-level healthcare research and to present some practical guidelines for its use. Two modeling approaches commonly used in healthcare research, system dynamics models and agent-based models, are presented and their applications in healthcare research are described. The three simulation paradigms, {M}onte {C}arlo simulation, continuous simulation, and discrete event simulation, are defined and the conditions for their use are stated. An epidemiology case study is presented to illustrate the use of M{\&}S in the research process.",
 isbn="978-3-030-26837-4",
 doi="10.1007/978-3-030-26837-4_6",
 url="https://doi.org/10.1007/978-3-030-26837-4_6"
