+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.1.7"
-release = "0.1.7"
+version = "0.1.8"
+release = "0.1.8"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
