+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.1.9"
-release = "0.1.9"
+version = "0.1.10"
+release = "0.1.10"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
