### 2024-12-01 #6970 @bc118
@jpotoff for notification.

@phibeck The suggests were added in this [PR 137](PR https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/pull/137).  These are present in the latest release 0.1.10, and we are currently waiting for the conda-forge bot to find it and then will will create a new conda-forge package and the Zenodo update.  We will update you when the new conda-forge package and the Zenodo update are completed.

The changes are currently [here](https://github.com/GOMC-WSU/MoSDeF-dihedral-fit/tree/main/paper).  Feel free to look at these now, if you wish.  Here are my summary answers to your suggestions/comments.
    • lines 6-11: in your affiliations, please spell out USA and UK
           Completed – as requested
                ◦ United States of America
                ◦ United Kingdom
    • line 38: "do not"
           Completed as requested
    • line 41: "are not"
           Completed as requested
    • line 42: "due to the different parameters" \
           Completed as requested
    • line 45: "allows users to quickly calculate"
           Completed  as requested
    • line 51: if I understand correctly, Gaussian16 .log is a file format, right? To distinguish file formats from regular text it can be useful to just mark them as .log I would probably write "provide a Gaussian16 .log file or Gaussian-like QM simulation file that covers..." XML can be either left as is or written as .xml too
           Completed  as requested – wrote .log, .xml, and .mol2 as suggested
    • line 52: "(typically between 0 and 360 degrees)
           Completed  as requested
    • line 53: Again, I would recommend `.mol2
           Completed  as requested – wrote .log, .xml, and .mol2 as suggested
    • line 56: remove hyphen in "(RB)-torsions" (and only introduce acronym if needed later, see line 98).
           Completed  as requested  on line 56 and 98– wrote .log, .xml, and .mol2 as suggested
    • line 63: MM acronym already defined above. Additionally, I would suggest "While many of these MM force field parameters such as bonds, angles and improper dihedrals (impropers) can be transferred between force fields, ..." I would also suggest to write "(often referred to as "impropers")" if that is correct.
           Completed  as requested both
                   ◦ Only using MM in line 63
                   ◦ rewriting as "(often referred to as "impropers")"
    • line 69: remove "absolutely"
           Completed  as requested
    • line 75: "there is a need for a"
           Completed  as requested
    • line 86: add an @ in front of "Jorgensen:1996" to render reference correctly
           Completed  as requested
    • lines 98f: I am not sure of the use of the equations in the paper. At present, they are neither referenced in the manuscript nor explained including all parameters, making them less helpful for the manuscript's intended audience (i.e., a diverse, non-specialist group).
           Completed  as requested  - To do this we added
                   ◦ Equation numbers to the OPLS dihedral, Ryckaert-Bellemans torsion, and the periodic dihedral forms
                   ◦ Added the equation reference in the 2nd paragraph of the Summary, with the reason some constants are set in the                            Ryckaert-Bellemans torsion, and the periodic dihedral forms, which reads “The MoSDeF-Dihedral-Fit software uses the
QM and MM data to produce the dihedral for the specific force field, fitting the
constants for the OPLS dihedral form (equation 1) and then analytically con-
verting them to the Ryckaert-Bellemans torsion (equation 2) and the periodic
dihedral forms (equation 3). This analytical conversion from the OPLS dihedral
form requires setting the specified parameters in the Ryckaert-Bellemans torsion
and periodic dihedral forms (see equations 2 and 3).”

    • line 109: "1-n" with n in italic font
           Completed  as requested
    • line 116: capitalize Monte Carlo
           Completed  as requested
    • line 120: capitalize Monte Carlo
           Completed  as requested
    • line 126: capitalize Molecular Simulation Design Framework
           Completed  as requested
    • line 129: "n-alkane" with n in italic font
           Completed  as requested
    • line 182: check correct capitalization of title
           Completed  as requested – Changed in .bib file to -
           “{U}eber die {A}nwendung des {S}atzes vom {V}irial in der kinetischen {T}heorie der {G}ase”

**_As a side note:_** when running the recommend Pandoc software to PDF this paper, it does show the correct capitalization without these changes for the references, despite most of them being pulled as a .bib from the publishers website.  Unless I missed it, JOSS may want to mention this in the JOSS documentation... need to manually ensure reference capitalization.

    • line 192: "n-alkanes" with n in italic font
Completed  as requested
