### 2024-12-02 #6970 @bc118
@phibeck @jpotoff Since the conda-forge bots are not finding it automatically, I manually pushed the conda-forge `mosdef-dihedral-fit`  package version `0.1.10`.  It is now on conda-forge and passing all the test.

@jpotoff Please update the `zenodo` `mosdef-dihedral-fit`  package version `0.1.10`. and provide the DOI and other needed info on here.
