### 2024-12-02 #6970 @bc118
@jpotoff For information.

@phibeck Hello!  This should all be done now and ready for your review.

The `mosdef-dihedral fit` software package (version` 0.1.10`) with our changes to your suggestions and comments is now on conda-forge and zendo (Zenodo DOI = `10.5281/zenodo.14254346`     and  link is  [here]((https://doi.org/10.5281/zenodo.14254346).

Note:   Conda-forge was not getting the packages with the automated bot as usual, so we had to do a few things to test and do manually; hence the version` 0.1.10`.
Please let us know if you are OK with the changes.
