### 2024-12-03 #6970 @phibeck
Hi @bc118 , @jpotoff, thank you for the update. This looks good to me and is ready to be handed over to the Track Editor-in-chief (@openjournals/pe-eics), who might also have some thoughts on your observation:

> **_As a side note:_** when running the recommend Pandoc software to PDF this paper, it does show the correct capitalization without these changes for the references, despite most of them being pulled as a .bib from the publishers website. Unless I missed it, JOSS may want to mention this in the JOSS documentation... need to manually ensure reference capitalization.
