### 2024-12-03 #6970 @xuanxu
It looks like the `Mayne:2013` entry is duplicated in the .bib file
