+++ paper/paper.bib
@@ -269,7 +269,8 @@ URL = {https://doi.org/10.1021/acs.jcim.2c01498},
 	title = "{M}o{SD}e{F}-{GOMC}: Python software for the creation of scientific workflows for the {M}onte {C}arlo simulation engine {GOMC}",
 	year = "2022",
 	publisher = "Github",
-	url = "https://github.com/GOMC-WSU/MoSDeF-GOMC"
+	url = "https://github.com/GOMC-WSU/MoSDeF-GOMC",
+  doi = {10.5281/zenodo.14266177}
 }
 
 # GOMC part 1
@@ -439,19 +440,6 @@ pages = {148},
 year = {2023},
 doi = {10.1038/s42004-023-00944-z},
 URL = {https://doi.org/10.1038/s42004-023-00944-z}
-}
-
-# CHARMM FFTK using vibrational spectra
-@article{Mayne:2013,
-author = {Christopher G. Mayne and Jan Saam and Klaus Schulten and Emad Tajkhorshid and James C. Gumbart},
-title = {Rapid parameterization of small molecules using the Force Field Toolkit},
-journal = {J Comput Chem.},
-volume = {34},
-issue = {32},
-pages = {2757-2770},
-year = {2013},
-doi = {10.1002/jcc.23422},
-URL = {https://doi.org/10.1002/jcc.23422}
 }
 
 # dihedral - bonded parameters of GROMOS
