### 2024-12-03 5ba04c1
removed duplicate citation
