+++ paper/paper.bib
@@ -245,7 +245,7 @@ year = {2020}
 	year        = "2023",
 	publisher   = "Github",
 	url         = {https://github.com/GOMC-WSU/MoSDeF-dihedral-fit},
-	doi = {10.5281/zenodo.14107384}
+	doi         =  {10.5281/zenodo.14107384}
 
 }
 
