### 2024-12-03 ecbdade
removed duplicate citation
