+++ paper/paper.bib
@@ -270,7 +270,7 @@ URL = {https://doi.org/10.1021/acs.jcim.2c01498},
 	year = "2022",
 	publisher = "Github",
 	url = "https://github.com/GOMC-WSU/MoSDeF-GOMC",
-  doi = {10.5281/zenodo.14266177}
+	doi = {10.5281/zenodo.14266177}
 }
 
 # GOMC part 1
