### 2024-12-03 6984fae
added References title and fixed one reference for the JOSS paper.
