### 2024-12-03 #6970 @bc118
@phibeck @jpotoff Upon review noticed that 1 reference needed corrected and that the Title "References" was missing.  I am correcting this in version 0.1.12 now, which will be available very soon
