### 2024-12-03 bfead16
Bump to version 0.1.12
