+++ docs/conf.py
@@ -157,8 +157,8 @@ copyright = "2023"
 # built documents.
 #
 
-version = "0.1.11"
-release = "0.1.11"
+version = "0.1.12"
+release = "0.1.12"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
