+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.1.11
+current_version = 0.1.12
 commit = True
 tag = True
 message = Bump to version {new_version}
