### 2024-12-04 #6970 @bc118
@jpotoff For information

@phibeck  Upon review noticed that 1 reference needed corrected (not printing correctly in the paper) and that the Title "References" was missing.  It is now updated to version `0.1.12` as a `GitHub`, `conda-forge`, and `Zenodo` packages. I have listed the relevant information below:

    zenodo DOI for version `0.1.12` = `10.5281/zenodo.14271075` and link is [here](https://doi.org/10.5281/zenodo.14271075).

It should be good now!
