### 2024-12-04 #6970 @GOMC-WSU
@phibeck The proof looks good to me.
