### 2024-12-05 #6970 @kyleniemeyer
Hi @jpotoff, in the paper, can the equations be moved to where they are actually mentioned in the text, rather than grouped together at the end (without the appropriate context)?

It also appears that they have some unnecessary linebreaks, and are missing appropriate `\text{}` commands around some subscripts and words ("where", "periodic", etc.). In addition, functions like `cos` should be expressed as `\cos` so they appear in the appropriate font and are not interpreted as a sequence of variables.
