### 2024-12-06 cc90ccf
additional editor suggestions added to paper, which are on the equations edits and moving the equations in to the main paper.
