### 2024-12-06 #6970 @bc118
@kyleniemeyer @jpotoff @phibeck

I have made the additional changes that were suggested by @kyleniemeyer above.

It is now updated to version `0.1.13` as `GitHub`, `conda-forge`, and `Zenodo` packages. I have listed the relevant information below:

zenodo DOI for version `0.1.13` = `10.5281/zenodo.14288503` and link is [here](https://doi.org/10.5281/zenodo.14288503).
