### 2024-12-06 #6970 @kyleniemeyer
@bc118 ah, changes to the paper only don't actually require updating the version / archives, if the software didn't change. But that's fine.
