### 2024-12-06 #6970 @kyleniemeyer
Congratulations @jpotoff @bc118 on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.

Many thanks to @naik-aakash and @RiesBen for reviewing this, and @phibeck for editing it.
